Crate core_simd

Portable SIMD module.

Structs

BitMask
A mask where each lane is represented by a single bit.

Mask8
Mask for vectors with LANES 8-bit elements.

Mask16
Mask for vectors with LANES 16-bit elements.

Mask32
Mask for vectors with LANES 32-bit elements.

Mask64
Mask for vectors with LANES 64-bit elements.

Mask128
Mask for vectors with LANES 128-bit elements.

MaskSize
Mask for vectors with LANES pointer-width elements.

SimdF32
A SIMD vector of containing LANES f32 values.

SimdF64
A SIMD vector of containing LANES f64 values.

SimdI8
A SIMD vector of containing LANES i8 values.

SimdI16
A SIMD vector of containing LANES i16 values.

SimdI32
A SIMD vector of containing LANES i32 values.

SimdI64
A SIMD vector of containing LANES i64 values.

SimdI128
A SIMD vector of containing LANES i128 values.

SimdIsize
A SIMD vector of containing LANES isize values.

SimdMask8
A mask equivalent to SimdI8, where all bits in the lane must be either set or unset.

SimdMask16
A mask equivalent to SimdI16, where all bits in the lane must be either set or unset.

SimdMask32
A mask equivalent to SimdI32, where all bits in the lane must be either set or unset.

SimdMask64
A mask equivalent to SimdI64, where all bits in the lane must be either set or unset.

SimdMask128
A mask equivalent to SimdI128, where all bits in the lane must be either set or unset.

SimdMaskSize
A mask equivalent to SimdIsize, where all bits in the lane must be either set or unset.

SimdU8
A SIMD vector of containing LANES u8 values.

SimdU16
A SIMD vector of containing LANES u16 values.

SimdU32
A SIMD vector of containing LANES u32 values.

SimdU64
A SIMD vector of containing LANES u64 values.

SimdU128
A SIMD vector of containing LANES u128 values.

SimdUsize
A SIMD vector of containing LANES usize values.

TryFromMaskError
The error type returned when converting an integer to a mask fails.

Traits

LanesAtMost64
Implemented for bitmask sizes that are supported by the implementation.

Type Definitions

f32x2
Vector of two f32 values

f32x4
Vector of four f32 values

f32x8
Vector of eight f32 values

f32x16
Vector of 16 f32 values

f64x2
Vector of two f64 values

f64x4
Vector of four f64 values

f64x8
Vector of eight f64 values

i8x8
Vector of eight i8 values

i8x16
Vector of 16 i8 values

i8x32
Vector of 32 i8 values

i8x64
Vector of 64 i8 values

i16x4
Vector of four i16 values

i16x8
Vector of eight i16 values

i16x16
Vector of 16 i16 values

i16x32
Vector of 32 i16 values

i32x2
Vector of two i32 values

i32x4
Vector of four i32 values

i32x8
Vector of eight i32 values

i32x16
Vector of 16 i32 values

i64x2
Vector of two i64 values

i64x4
Vector of four i64 values

i64x8
Vector of eight i64 values

i128x2
Vector of two i128 values

i128x4
Vector of four i128 values

isizex2
Vector of two isize values

isizex4
Vector of four isize values

isizex8
Vector of eight isize values

mask8x8
Vector of eight 8-bit masks

mask8x16
Vector of 16 8-bit masks

mask8x32
Vector of 32 8-bit masks

mask8x64
Vector of 16 8-bit masks

mask16x4
Vector of four 16-bit masks

mask16x8
Vector of eight 16-bit masks

mask16x16
Vector of 16 16-bit masks

mask16x32
Vector of 32 16-bit masks

mask32x2
Vector of two 32-bit masks

mask32x4
Vector of four 32-bit masks

mask32x8
Vector of eight 32-bit masks

mask32x16
Vector of 16 32-bit masks

mask64x2
Vector of two 64-bit masks

mask64x4
Vector of four 64-bit masks

mask64x8
Vector of eight 64-bit masks

mask128x2
Vector of two 128-bit masks

mask128x4
Vector of four 128-bit masks

masksizex2
Vector of two pointer-width masks

masksizex4
Vector of four pointer-width masks

masksizex8
Vector of eight pointer-width masks

u8x8
Vector of eight u8 values

u8x16
Vector of 16 u8 values

u8x32
Vector of 32 u8 values

u8x64
Vector of 64 u8 values

u16x4
Vector of four u16 values

u16x8
Vector of eight u16 values

u16x16
Vector of 16 u16 values

u16x32
Vector of 32 u16 values

u32x2
Vector of two u32 values

u32x4
Vector of four u32 values

u32x8
Vector of eight u32 values

u32x16
Vector of 16 u32 values

u64x2
Vector of two u64 values

u64x4
Vector of four u64 values

u64x8
Vector of eight u64 values

u128x2
Vector of two u128 values

u128x4
Vector of four u128 values

usizex2
Vector of two usize values

usizex4
Vector of four usize values

usizex8
Vector of eight usize values