Module core_simd::full_masks

Masks that take up full SIMD vector registers.

Structs

SimdI8Mask
A mask equivalent to SimdI8, where all bits in the lane must be either set or unset.

SimdI16Mask
A mask equivalent to SimdI16, where all bits in the lane must be either set or unset.

SimdI32Mask
A mask equivalent to SimdI32, where all bits in the lane must be either set or unset.

SimdI64Mask
A mask equivalent to SimdI64, where all bits in the lane must be either set or unset.

SimdI128Mask
A mask equivalent to SimdI128, where all bits in the lane must be either set or unset.

SimdIsizeMask
A mask equivalent to SimdIsize, where all bits in the lane must be either set or unset.

TryFromMaskError
The error type returned when converting an integer to a mask fails.