Struct core_simd::simd::Simd

#[repr(simd)]
pub struct Simd<T, const LANES: usize>(_)
 where
    T: SimdElement,
    LaneCount<LANES>: SupportedLaneCount;

🔬 This is a nightly-only experimental API. (portable_simd)

A SIMD vector of LANES elements of type T. Simd<T, N> has the same shape as [T; N], but operates like T.

Two vectors of the same type and length will, by convention, support the operators (+, *, etc.) that T does. These take the lanes at each index on the left-hand side and right-hand side, perform the operation, and return the result in the same lane in a vector of equal size. For a given operator, this is equivalent to zipping the two arrays together and mapping the operator over each lane.

let a0: [i32; 4] = [-2, 0, 2, 4];
let a1 = [10, 9, 8, 7];
let zm_add = a0.zip(a1).map(|(lhs, rhs)| lhs + rhs);
let zm_mul = a0.zip(a1).map(|(lhs, rhs)| lhs * rhs);

// `Simd<T, N>` implements `From<[T; N]>
let (v0, v1) = (Simd::from(a0), Simd::from(a1));
// Which means arrays implement `Into<Simd<T, N>>`.
assert_eq!(v0 + v1, zm_add.into());
assert_eq!(v0 * v1, zm_mul.into());

Simd with integers has the quirk that these operations are also inherently wrapping, as if T was Wrapping<T>. Thus, Simd does not implement wrapping_add, because that is the default behavior. This means there is no warning on overflows, even in “debug” builds. For most applications where Simd is appropriate, it is “not a bug” to wrap, and even “debug builds” are unlikely to tolerate the loss of performance. You may want to consider using explicitly checked arithmetic if such is required. Division by zero still causes a panic, so you may want to consider using floating point numbers if that is unacceptable.

Implementations

impl<const LANES: usize> Simd<i8, LANES> where
LaneCount<LANES>: SupportedLaneCount,

pub fn horizontal_sum(self) -> i8

🔬 This is a nightly-only experimental API. (portable_simd)

Horizontal wrapping add. Returns the sum of the lanes of the vector, with wrapping addition.

pub fn horizontal_product(self) -> i8

🔬 This is a nightly-only experimental API. (portable_simd)

Horizontal wrapping multiply. Returns the product of the lanes of the vector, with wrapping multiplication.

pub fn horizontal_max(self) -> i8

🔬 This is a nightly-only experimental API. (portable_simd)

Horizontal maximum. Returns the maximum lane in the vector.

pub fn horizontal_min(self) -> i8

🔬 This is a nightly-only experimental API. (portable_simd)

Horizontal minimum. Returns the minimum lane in the vector.

impl<const LANES: usize> Simd<i16, LANES> where
LaneCount<LANES>: SupportedLaneCount,

pub fn horizontal_sum(self) -> i16

🔬 This is a nightly-only experimental API. (portable_simd)

Horizontal wrapping add. Returns the sum of the lanes of the vector, with wrapping addition.

pub fn horizontal_product(self) -> i16

🔬 This is a nightly-only experimental API. (portable_simd)

Horizontal wrapping multiply. Returns the product of the lanes of the vector, with wrapping multiplication.

pub fn horizontal_max(self) -> i16

🔬 This is a nightly-only experimental API. (portable_simd)

Horizontal maximum. Returns the maximum lane in the vector.

pub fn horizontal_min(self) -> i16

🔬 This is a nightly-only experimental API. (portable_simd)

Horizontal minimum. Returns the minimum lane in the vector.

impl<const LANES: usize> Simd<i32, LANES> where
LaneCount<LANES>: SupportedLaneCount,

pub fn horizontal_sum(self) -> i32

🔬 This is a nightly-only experimental API. (portable_simd)

Horizontal wrapping add. Returns the sum of the lanes of the vector, with wrapping addition.

pub fn horizontal_product(self) -> i32

🔬 This is a nightly-only experimental API. (portable_simd)

Horizontal wrapping multiply. Returns the product of the lanes of the vector, with wrapping multiplication.

pub fn horizontal_max(self) -> i32

🔬 This is a nightly-only experimental API. (portable_simd)

Horizontal maximum. Returns the maximum lane in the vector.

pub fn horizontal_min(self) -> i32

🔬 This is a nightly-only experimental API. (portable_simd)

Horizontal minimum. Returns the minimum lane in the vector.

impl<const LANES: usize> Simd<i64, LANES> where
LaneCount<LANES>: SupportedLaneCount,

pub fn horizontal_sum(self) -> i64

🔬 This is a nightly-only experimental API. (portable_simd)

Horizontal wrapping add. Returns the sum of the lanes of the vector, with wrapping addition.

pub fn horizontal_product(self) -> i64

🔬 This is a nightly-only experimental API. (portable_simd)

Horizontal wrapping multiply. Returns the product of the lanes of the vector, with wrapping multiplication.

pub fn horizontal_max(self) -> i64

🔬 This is a nightly-only experimental API. (portable_simd)

Horizontal maximum. Returns the maximum lane in the vector.

pub fn horizontal_min(self) -> i64

🔬 This is a nightly-only experimental API. (portable_simd)

Horizontal minimum. Returns the minimum lane in the vector.

impl<const LANES: usize> Simd<isize, LANES> where
LaneCount<LANES>: SupportedLaneCount,

pub fn horizontal_sum(self) -> isize

🔬 This is a nightly-only experimental API. (portable_simd)

Horizontal wrapping add. Returns the sum of the lanes of the vector, with wrapping addition.

pub fn horizontal_product(self) -> isize

🔬 This is a nightly-only experimental API. (portable_simd)

Horizontal wrapping multiply. Returns the product of the lanes of the vector, with wrapping multiplication.

pub fn horizontal_max(self) -> isize

🔬 This is a nightly-only experimental API. (portable_simd)

Horizontal maximum. Returns the maximum lane in the vector.

pub fn horizontal_min(self) -> isize

🔬 This is a nightly-only experimental API. (portable_simd)

Horizontal minimum. Returns the minimum lane in the vector.

impl<const LANES: usize> Simd<u8, LANES> where
LaneCount<LANES>: SupportedLaneCount,

pub fn horizontal_sum(self) -> u8

🔬 This is a nightly-only experimental API. (portable_simd)

Horizontal wrapping add. Returns the sum of the lanes of the vector, with wrapping addition.

pub fn horizontal_product(self) -> u8

🔬 This is a nightly-only experimental API. (portable_simd)

Horizontal wrapping multiply. Returns the product of the lanes of the vector, with wrapping multiplication.

pub fn horizontal_max(self) -> u8

🔬 This is a nightly-only experimental API. (portable_simd)

Horizontal maximum. Returns the maximum lane in the vector.

pub fn horizontal_min(self) -> u8

🔬 This is a nightly-only experimental API. (portable_simd)

Horizontal minimum. Returns the minimum lane in the vector.

impl<const LANES: usize> Simd<u16, LANES> where
LaneCount<LANES>: SupportedLaneCount,

pub fn horizontal_sum(self) -> u16

🔬 This is a nightly-only experimental API. (portable_simd)

Horizontal wrapping add. Returns the sum of the lanes of the vector, with wrapping addition.

pub fn horizontal_product(self) -> u16

🔬 This is a nightly-only experimental API. (portable_simd)

Horizontal wrapping multiply. Returns the product of the lanes of the vector, with wrapping multiplication.

pub fn horizontal_max(self) -> u16

🔬 This is a nightly-only experimental API. (portable_simd)

Horizontal maximum. Returns the maximum lane in the vector.

pub fn horizontal_min(self) -> u16

🔬 This is a nightly-only experimental API. (portable_simd)

Horizontal minimum. Returns the minimum lane in the vector.

impl<const LANES: usize> Simd<u32, LANES> where
LaneCount<LANES>: SupportedLaneCount,

pub fn horizontal_sum(self) -> u32

🔬 This is a nightly-only experimental API. (portable_simd)

Horizontal wrapping add. Returns the sum of the lanes of the vector, with wrapping addition.

pub fn horizontal_product(self) -> u32

🔬 This is a nightly-only experimental API. (portable_simd)

Horizontal wrapping multiply. Returns the product of the lanes of the vector, with wrapping multiplication.

pub fn horizontal_max(self) -> u32

🔬 This is a nightly-only experimental API. (portable_simd)

Horizontal maximum. Returns the maximum lane in the vector.

pub fn horizontal_min(self) -> u32

🔬 This is a nightly-only experimental API. (portable_simd)

Horizontal minimum. Returns the minimum lane in the vector.

impl<const LANES: usize> Simd<u64, LANES> where
LaneCount<LANES>: SupportedLaneCount,

pub fn horizontal_sum(self) -> u64

🔬 This is a nightly-only experimental API. (portable_simd)

Horizontal wrapping add. Returns the sum of the lanes of the vector, with wrapping addition.

pub fn horizontal_product(self) -> u64

🔬 This is a nightly-only experimental API. (portable_simd)

Horizontal wrapping multiply. Returns the product of the lanes of the vector, with wrapping multiplication.

pub fn horizontal_max(self) -> u64

🔬 This is a nightly-only experimental API. (portable_simd)

Horizontal maximum. Returns the maximum lane in the vector.

pub fn horizontal_min(self) -> u64

🔬 This is a nightly-only experimental API. (portable_simd)

Horizontal minimum. Returns the minimum lane in the vector.

impl<const LANES: usize> Simd<usize, LANES> where
LaneCount<LANES>: SupportedLaneCount,

pub fn horizontal_sum(self) -> usize

🔬 This is a nightly-only experimental API. (portable_simd)

Horizontal wrapping add. Returns the sum of the lanes of the vector, with wrapping addition.

pub fn horizontal_product(self) -> usize

🔬 This is a nightly-only experimental API. (portable_simd)

Horizontal wrapping multiply. Returns the product of the lanes of the vector, with wrapping multiplication.

pub fn horizontal_max(self) -> usize

🔬 This is a nightly-only experimental API. (portable_simd)

Horizontal maximum. Returns the maximum lane in the vector.

pub fn horizontal_min(self) -> usize

🔬 This is a nightly-only experimental API. (portable_simd)

Horizontal minimum. Returns the minimum lane in the vector.

impl<const LANES: usize> Simd<f32, LANES> where
LaneCount<LANES>: SupportedLaneCount,

pub fn horizontal_sum(self) -> f32

🔬 This is a nightly-only experimental API. (portable_simd)

Horizontal add. Returns the sum of the lanes of the vector.

pub fn horizontal_product(self) -> f32

🔬 This is a nightly-only experimental API. (portable_simd)

Horizontal multiply. Returns the product of the lanes of the vector.

pub fn horizontal_max(self) -> f32

🔬 This is a nightly-only experimental API. (portable_simd)

Horizontal maximum. Returns the maximum lane in the vector.

Returns values based on equality, so a vector containing both 0. and -0. may return either. This function will not return NaN unless all lanes are NaN.

pub fn horizontal_min(self) -> f32

🔬 This is a nightly-only experimental API. (portable_simd)

Horizontal minimum. Returns the minimum lane in the vector.

Returns values based on equality, so a vector containing both 0. and -0. may return either. This function will not return NaN unless all lanes are NaN.

impl<const LANES: usize> Simd<f64, LANES> where
LaneCount<LANES>: SupportedLaneCount,

pub fn horizontal_sum(self) -> f64

🔬 This is a nightly-only experimental API. (portable_simd)

Horizontal add. Returns the sum of the lanes of the vector.

pub fn horizontal_product(self) -> f64

🔬 This is a nightly-only experimental API. (portable_simd)

Horizontal multiply. Returns the product of the lanes of the vector.

pub fn horizontal_max(self) -> f64

🔬 This is a nightly-only experimental API. (portable_simd)

Horizontal maximum. Returns the maximum lane in the vector.

Returns values based on equality, so a vector containing both 0. and -0. may return either. This function will not return NaN unless all lanes are NaN.

pub fn horizontal_min(self) -> f64

🔬 This is a nightly-only experimental API. (portable_simd)

Horizontal minimum. Returns the minimum lane in the vector.

Returns values based on equality, so a vector containing both 0. and -0. may return either. This function will not return NaN unless all lanes are NaN.

impl<T, const LANES: usize> Simd<T, LANES> where
Self: BitAnd<Self, Output = Self>,
T: SimdElement + BitAnd<T, Output = T>,
LaneCount<LANES>: SupportedLaneCount,

pub fn horizontal_and(self) -> T

🔬 This is a nightly-only experimental API. (portable_simd)

Horizontal bitwise “and”. Returns the cumulative bitwise “and” across the lanes of the vector.

impl<T, const LANES: usize> Simd<T, LANES> where
Self: BitOr<Self, Output = Self>,
T: SimdElement + BitOr<T, Output = T>,
LaneCount<LANES>: SupportedLaneCount,

pub fn horizontal_or(self) -> T

🔬 This is a nightly-only experimental API. (portable_simd)

Horizontal bitwise “or”. Returns the cumulative bitwise “or” across the lanes of the vector.

impl<T, const LANES: usize> Simd<T, LANES> where
Self: BitXor<Self, Output = Self>,
T: SimdElement + BitXor<T, Output = T>,
LaneCount<LANES>: SupportedLaneCount,

pub fn horizontal_xor(self) -> T

🔬 This is a nightly-only experimental API. (portable_simd)

Horizontal bitwise “xor”. Returns the cumulative bitwise “xor” across the lanes of the vector.

impl<T, const LANES: usize> Simd<T, LANES> where
T: SimdElement,
LaneCount<LANES>: SupportedLaneCount,

pub fn reverse(self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Reverse the order of the lanes in the vector.

pub fn rotate_lanes_left<const OFFSET: usize>(self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Rotates the vector such that the first OFFSET elements of the slice move to the end while the last LANES - OFFSET elements move to the front. After calling rotate_lanes_left, the element previously in lane OFFSET will become the first element in the slice.

pub fn rotate_lanes_right<const OFFSET: usize>(self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Rotates the vector such that the first LANES - OFFSET elements of the vector move to the end while the last OFFSET elements move to the front. After calling rotate_lanes_right, the element previously at index LANES - OFFSET will become the first element in the slice.

pub fn interleave(self, other: Self) -> (Self, Self)

🔬 This is a nightly-only experimental API. (portable_simd)

Interleave two vectors.

Produces two vectors with lanes taken alternately from self and other.

The first result contains the first LANES / 2 lanes from self and other, alternating, starting with the first lane of self.

The second result contains the last LANES / 2 lanes from self and other, alternating, starting with the lane LANES / 2 from the start of self.

#![feature(portable_simd)]
let a = Simd::from_array([0, 1, 2, 3]);
let b = Simd::from_array([4, 5, 6, 7]);
let (x, y) = a.interleave(b);
assert_eq!(x.to_array(), [0, 4, 1, 5]);
assert_eq!(y.to_array(), [2, 6, 3, 7]);

pub fn deinterleave(self, other: Self) -> (Self, Self)

🔬 This is a nightly-only experimental API. (portable_simd)

Deinterleave two vectors.

The first result takes every other lane of self and then other, starting with the first lane.

The second result takes every other lane of self and then other, starting with the second lane.

#![feature(portable_simd)]
let a = Simd::from_array([0, 4, 1, 5]);
let b = Simd::from_array([2, 6, 3, 7]);
let (x, y) = a.deinterleave(b);
assert_eq!(x.to_array(), [0, 1, 2, 3]);
assert_eq!(y.to_array(), [4, 5, 6, 7]);

impl<const LANES: usize> Simd<u8, LANES> where
LaneCount<LANES>: SupportedLaneCount,
LaneCount<{ { 1 * LANES } }>: SupportedLaneCount,

pub fn to_ne_bytes(self) -> Simd<u8, { { 1 * LANES } }>

🔬 This is a nightly-only experimental API. (portable_simd)

Return the memory representation of this integer as a byte array in native byte order.

pub fn from_ne_bytes(bytes: Simd<u8, { { 1 * LANES } }>) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Create a native endian integer value from its memory representation as a byte array in native endianness.

impl<const LANES: usize> Simd<u16, LANES> where
LaneCount<LANES>: SupportedLaneCount,
LaneCount<{ { 2 * LANES } }>: SupportedLaneCount,

pub fn to_ne_bytes(self) -> Simd<u8, { { 2 * LANES } }>

🔬 This is a nightly-only experimental API. (portable_simd)

Return the memory representation of this integer as a byte array in native byte order.

pub fn from_ne_bytes(bytes: Simd<u8, { { 2 * LANES } }>) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Create a native endian integer value from its memory representation as a byte array in native endianness.

impl<const LANES: usize> Simd<u32, LANES> where
LaneCount<LANES>: SupportedLaneCount,
LaneCount<{ { 4 * LANES } }>: SupportedLaneCount,

pub fn to_ne_bytes(self) -> Simd<u8, { { 4 * LANES } }>

🔬 This is a nightly-only experimental API. (portable_simd)

Return the memory representation of this integer as a byte array in native byte order.

pub fn from_ne_bytes(bytes: Simd<u8, { { 4 * LANES } }>) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Create a native endian integer value from its memory representation as a byte array in native endianness.

impl<const LANES: usize> Simd<u64, LANES> where
LaneCount<LANES>: SupportedLaneCount,
LaneCount<{ { 8 * LANES } }>: SupportedLaneCount,

pub fn to_ne_bytes(self) -> Simd<u8, { { 8 * LANES } }>

🔬 This is a nightly-only experimental API. (portable_simd)

Return the memory representation of this integer as a byte array in native byte order.

pub fn from_ne_bytes(bytes: Simd<u8, { { 8 * LANES } }>) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Create a native endian integer value from its memory representation as a byte array in native endianness.

impl<const LANES: usize> Simd<usize, LANES> where
LaneCount<LANES>: SupportedLaneCount,
LaneCount<{ { 8 * LANES } }>: SupportedLaneCount,

pub fn to_ne_bytes(self) -> Simd<u8, { { 8 * LANES } }>

🔬 This is a nightly-only experimental API. (portable_simd)

Return the memory representation of this integer as a byte array in native byte order.

pub fn from_ne_bytes(bytes: Simd<u8, { { 8 * LANES } }>) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Create a native endian integer value from its memory representation as a byte array in native endianness.

impl<const LANES: usize> Simd<i8, LANES> where
LaneCount<LANES>: SupportedLaneCount,
LaneCount<{ { 1 * LANES } }>: SupportedLaneCount,

pub fn to_ne_bytes(self) -> Simd<u8, { { 1 * LANES } }>

🔬 This is a nightly-only experimental API. (portable_simd)

Return the memory representation of this integer as a byte array in native byte order.

pub fn from_ne_bytes(bytes: Simd<u8, { { 1 * LANES } }>) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Create a native endian integer value from its memory representation as a byte array in native endianness.

impl<const LANES: usize> Simd<i16, LANES> where
LaneCount<LANES>: SupportedLaneCount,
LaneCount<{ { 2 * LANES } }>: SupportedLaneCount,

pub fn to_ne_bytes(self) -> Simd<u8, { { 2 * LANES } }>

🔬 This is a nightly-only experimental API. (portable_simd)

Return the memory representation of this integer as a byte array in native byte order.

pub fn from_ne_bytes(bytes: Simd<u8, { { 2 * LANES } }>) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Create a native endian integer value from its memory representation as a byte array in native endianness.

impl<const LANES: usize> Simd<i32, LANES> where
LaneCount<LANES>: SupportedLaneCount,
LaneCount<{ { 4 * LANES } }>: SupportedLaneCount,

pub fn to_ne_bytes(self) -> Simd<u8, { { 4 * LANES } }>

🔬 This is a nightly-only experimental API. (portable_simd)

Return the memory representation of this integer as a byte array in native byte order.

pub fn from_ne_bytes(bytes: Simd<u8, { { 4 * LANES } }>) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Create a native endian integer value from its memory representation as a byte array in native endianness.

impl<const LANES: usize> Simd<i64, LANES> where
LaneCount<LANES>: SupportedLaneCount,
LaneCount<{ { 8 * LANES } }>: SupportedLaneCount,

pub fn to_ne_bytes(self) -> Simd<u8, { { 8 * LANES } }>

🔬 This is a nightly-only experimental API. (portable_simd)

Return the memory representation of this integer as a byte array in native byte order.

pub fn from_ne_bytes(bytes: Simd<u8, { { 8 * LANES } }>) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Create a native endian integer value from its memory representation as a byte array in native endianness.

impl<const LANES: usize> Simd<isize, LANES> where
LaneCount<LANES>: SupportedLaneCount,
LaneCount<{ { 8 * LANES } }>: SupportedLaneCount,

pub fn to_ne_bytes(self) -> Simd<u8, { { 8 * LANES } }>

🔬 This is a nightly-only experimental API. (portable_simd)

Return the memory representation of this integer as a byte array in native byte order.

pub fn from_ne_bytes(bytes: Simd<u8, { { 8 * LANES } }>) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Create a native endian integer value from its memory representation as a byte array in native endianness.

impl<T, const LANES: usize> Simd<T, LANES> where
T: SimdElement + PartialEq,
LaneCount<LANES>: SupportedLaneCount,

pub fn lanes_eq(self, other: Self) -> Mask<T::Mask, LANES>

🔬 This is a nightly-only experimental API. (portable_simd)

Test if each lane is equal to the corresponding lane in other.

pub fn lanes_ne(self, other: Self) -> Mask<T::Mask, LANES>

🔬 This is a nightly-only experimental API. (portable_simd)

Test if each lane is not equal to the corresponding lane in other.

impl<T, const LANES: usize> Simd<T, LANES> where
T: SimdElement + PartialOrd,
LaneCount<LANES>: SupportedLaneCount,

pub fn lanes_lt(self, other: Self) -> Mask<T::Mask, LANES>

🔬 This is a nightly-only experimental API. (portable_simd)

Test if each lane is less than the corresponding lane in other.

pub fn lanes_gt(self, other: Self) -> Mask<T::Mask, LANES>

🔬 This is a nightly-only experimental API. (portable_simd)

Test if each lane is greater than the corresponding lane in other.

pub fn lanes_le(self, other: Self) -> Mask<T::Mask, LANES>

🔬 This is a nightly-only experimental API. (portable_simd)

Test if each lane is less than or equal to the corresponding lane in other.

pub fn lanes_ge(self, other: Self) -> Mask<T::Mask, LANES>

🔬 This is a nightly-only experimental API. (portable_simd)

Test if each lane is greater than or equal to the corresponding lane in other.

impl<const LANES: usize> Simd<u8, LANES> where
LaneCount<LANES>: SupportedLaneCount,

pub fn saturating_add(self, second: Self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Lanewise saturating add.

Examples

let x = Simd::from_array([2, 1, 0, MAX]);
let max = Simd::splat(MAX);
let unsat = x + max;
let sat = x.saturating_add(max);
assert_eq!(unsat, Simd::from_array([1, 0, MAX, MAX - 1]));
assert_eq!(sat, max);

pub fn saturating_sub(self, second: Self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Lanewise saturating subtract.

Examples

let x = Simd::from_array([2, 1, 0, MAX]);
let max = Simd::splat(MAX);
let unsat = x - max;
let sat = x.saturating_sub(max);
assert_eq!(unsat, Simd::from_array([3, 2, 1, 0]));
assert_eq!(sat, Simd::splat(0));

impl<const LANES: usize> Simd<u16, LANES> where
LaneCount<LANES>: SupportedLaneCount,

pub fn saturating_add(self, second: Self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Lanewise saturating add.

Examples

let x = Simd::from_array([2, 1, 0, MAX]);
let max = Simd::splat(MAX);
let unsat = x + max;
let sat = x.saturating_add(max);
assert_eq!(unsat, Simd::from_array([1, 0, MAX, MAX - 1]));
assert_eq!(sat, max);

pub fn saturating_sub(self, second: Self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Lanewise saturating subtract.

Examples

let x = Simd::from_array([2, 1, 0, MAX]);
let max = Simd::splat(MAX);
let unsat = x - max;
let sat = x.saturating_sub(max);
assert_eq!(unsat, Simd::from_array([3, 2, 1, 0]));
assert_eq!(sat, Simd::splat(0));

impl<const LANES: usize> Simd<u32, LANES> where
LaneCount<LANES>: SupportedLaneCount,

pub fn saturating_add(self, second: Self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Lanewise saturating add.

Examples

let x = Simd::from_array([2, 1, 0, MAX]);
let max = Simd::splat(MAX);
let unsat = x + max;
let sat = x.saturating_add(max);
assert_eq!(unsat, Simd::from_array([1, 0, MAX, MAX - 1]));
assert_eq!(sat, max);

pub fn saturating_sub(self, second: Self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Lanewise saturating subtract.

Examples

let x = Simd::from_array([2, 1, 0, MAX]);
let max = Simd::splat(MAX);
let unsat = x - max;
let sat = x.saturating_sub(max);
assert_eq!(unsat, Simd::from_array([3, 2, 1, 0]));
assert_eq!(sat, Simd::splat(0));

impl<const LANES: usize> Simd<u64, LANES> where
LaneCount<LANES>: SupportedLaneCount,

pub fn saturating_add(self, second: Self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Lanewise saturating add.

Examples

let x = Simd::from_array([2, 1, 0, MAX]);
let max = Simd::splat(MAX);
let unsat = x + max;
let sat = x.saturating_add(max);
assert_eq!(unsat, Simd::from_array([1, 0, MAX, MAX - 1]));
assert_eq!(sat, max);

pub fn saturating_sub(self, second: Self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Lanewise saturating subtract.

Examples

let x = Simd::from_array([2, 1, 0, MAX]);
let max = Simd::splat(MAX);
let unsat = x - max;
let sat = x.saturating_sub(max);
assert_eq!(unsat, Simd::from_array([3, 2, 1, 0]));
assert_eq!(sat, Simd::splat(0));

impl<const LANES: usize> Simd<usize, LANES> where
LaneCount<LANES>: SupportedLaneCount,

pub fn saturating_add(self, second: Self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Lanewise saturating add.

Examples

let x = Simd::from_array([2, 1, 0, MAX]);
let max = Simd::splat(MAX);
let unsat = x + max;
let sat = x.saturating_add(max);
assert_eq!(unsat, Simd::from_array([1, 0, MAX, MAX - 1]));
assert_eq!(sat, max);

pub fn saturating_sub(self, second: Self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Lanewise saturating subtract.

Examples

let x = Simd::from_array([2, 1, 0, MAX]);
let max = Simd::splat(MAX);
let unsat = x - max;
let sat = x.saturating_sub(max);
assert_eq!(unsat, Simd::from_array([3, 2, 1, 0]));
assert_eq!(sat, Simd::splat(0));

impl<const LANES: usize> Simd<i8, LANES> where
LaneCount<LANES>: SupportedLaneCount,

pub fn saturating_add(self, second: Self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Lanewise saturating add.

Examples

let x = Simd::from_array([MIN, 0, 1, MAX]);
let max = Simd::splat(MAX);
let unsat = x + max;
let sat = x.saturating_add(max);
assert_eq!(unsat, Simd::from_array([-1, MAX, MIN, -2]));
assert_eq!(sat, Simd::from_array([-1, MAX, MAX, MAX]));

pub fn saturating_sub(self, second: Self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Lanewise saturating subtract.

Examples

let x = Simd::from_array([MIN, -2, -1, MAX]);
let max = Simd::splat(MAX);
let unsat = x - max;
let sat = x.saturating_sub(max);
assert_eq!(unsat, Simd::from_array([1, MAX, MIN, 0]));
assert_eq!(sat, Simd::from_array([MIN, MIN, MIN, 0]));

pub fn abs(self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Lanewise absolute value, implemented in Rust. Every lane becomes its absolute value.

Examples

let xs = Simd::from_array([MIN, MIN +1, -5, 0]);
assert_eq!(xs.abs(), Simd::from_array([MIN, MAX, 5, 0]));

pub fn saturating_abs(self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Lanewise saturating absolute value, implemented in Rust. As abs(), except the MIN value becomes MAX instead of itself.

Examples

let xs = Simd::from_array([MIN, -2, 0, 3]);
let unsat = xs.abs();
let sat = xs.saturating_abs();
assert_eq!(unsat, Simd::from_array([MIN, 2, 0, 3]));
assert_eq!(sat, Simd::from_array([MAX, 2, 0, 3]));

pub fn saturating_neg(self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Lanewise saturating negation, implemented in Rust. As neg(), except the MIN value becomes MAX instead of itself.

Examples

let x = Simd::from_array([MIN, -2, 3, MAX]);
let unsat = -x;
let sat = x.saturating_neg();
assert_eq!(unsat, Simd::from_array([MIN, 2, -3, MIN + 1]));
assert_eq!(sat, Simd::from_array([MAX, 2, -3, MIN + 1]));

impl<const LANES: usize> Simd<i16, LANES> where
LaneCount<LANES>: SupportedLaneCount,

pub fn saturating_add(self, second: Self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Lanewise saturating add.

Examples

let x = Simd::from_array([MIN, 0, 1, MAX]);
let max = Simd::splat(MAX);
let unsat = x + max;
let sat = x.saturating_add(max);
assert_eq!(unsat, Simd::from_array([-1, MAX, MIN, -2]));
assert_eq!(sat, Simd::from_array([-1, MAX, MAX, MAX]));

pub fn saturating_sub(self, second: Self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Lanewise saturating subtract.

Examples

let x = Simd::from_array([MIN, -2, -1, MAX]);
let max = Simd::splat(MAX);
let unsat = x - max;
let sat = x.saturating_sub(max);
assert_eq!(unsat, Simd::from_array([1, MAX, MIN, 0]));
assert_eq!(sat, Simd::from_array([MIN, MIN, MIN, 0]));

pub fn abs(self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Lanewise absolute value, implemented in Rust. Every lane becomes its absolute value.

Examples

let xs = Simd::from_array([MIN, MIN +1, -5, 0]);
assert_eq!(xs.abs(), Simd::from_array([MIN, MAX, 5, 0]));

pub fn saturating_abs(self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Lanewise saturating absolute value, implemented in Rust. As abs(), except the MIN value becomes MAX instead of itself.

Examples

let xs = Simd::from_array([MIN, -2, 0, 3]);
let unsat = xs.abs();
let sat = xs.saturating_abs();
assert_eq!(unsat, Simd::from_array([MIN, 2, 0, 3]));
assert_eq!(sat, Simd::from_array([MAX, 2, 0, 3]));

pub fn saturating_neg(self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Lanewise saturating negation, implemented in Rust. As neg(), except the MIN value becomes MAX instead of itself.

Examples

let x = Simd::from_array([MIN, -2, 3, MAX]);
let unsat = -x;
let sat = x.saturating_neg();
assert_eq!(unsat, Simd::from_array([MIN, 2, -3, MIN + 1]));
assert_eq!(sat, Simd::from_array([MAX, 2, -3, MIN + 1]));

impl<const LANES: usize> Simd<i32, LANES> where
LaneCount<LANES>: SupportedLaneCount,

pub fn saturating_add(self, second: Self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Lanewise saturating add.

Examples

let x = Simd::from_array([MIN, 0, 1, MAX]);
let max = Simd::splat(MAX);
let unsat = x + max;
let sat = x.saturating_add(max);
assert_eq!(unsat, Simd::from_array([-1, MAX, MIN, -2]));
assert_eq!(sat, Simd::from_array([-1, MAX, MAX, MAX]));

pub fn saturating_sub(self, second: Self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Lanewise saturating subtract.

Examples

let x = Simd::from_array([MIN, -2, -1, MAX]);
let max = Simd::splat(MAX);
let unsat = x - max;
let sat = x.saturating_sub(max);
assert_eq!(unsat, Simd::from_array([1, MAX, MIN, 0]));
assert_eq!(sat, Simd::from_array([MIN, MIN, MIN, 0]));

pub fn abs(self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Lanewise absolute value, implemented in Rust. Every lane becomes its absolute value.

Examples

let xs = Simd::from_array([MIN, MIN +1, -5, 0]);
assert_eq!(xs.abs(), Simd::from_array([MIN, MAX, 5, 0]));

pub fn saturating_abs(self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Lanewise saturating absolute value, implemented in Rust. As abs(), except the MIN value becomes MAX instead of itself.

Examples

let xs = Simd::from_array([MIN, -2, 0, 3]);
let unsat = xs.abs();
let sat = xs.saturating_abs();
assert_eq!(unsat, Simd::from_array([MIN, 2, 0, 3]));
assert_eq!(sat, Simd::from_array([MAX, 2, 0, 3]));

pub fn saturating_neg(self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Lanewise saturating negation, implemented in Rust. As neg(), except the MIN value becomes MAX instead of itself.

Examples

let x = Simd::from_array([MIN, -2, 3, MAX]);
let unsat = -x;
let sat = x.saturating_neg();
assert_eq!(unsat, Simd::from_array([MIN, 2, -3, MIN + 1]));
assert_eq!(sat, Simd::from_array([MAX, 2, -3, MIN + 1]));

impl<const LANES: usize> Simd<i64, LANES> where
LaneCount<LANES>: SupportedLaneCount,

pub fn saturating_add(self, second: Self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Lanewise saturating add.

Examples

let x = Simd::from_array([MIN, 0, 1, MAX]);
let max = Simd::splat(MAX);
let unsat = x + max;
let sat = x.saturating_add(max);
assert_eq!(unsat, Simd::from_array([-1, MAX, MIN, -2]));
assert_eq!(sat, Simd::from_array([-1, MAX, MAX, MAX]));

pub fn saturating_sub(self, second: Self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Lanewise saturating subtract.

Examples

let x = Simd::from_array([MIN, -2, -1, MAX]);
let max = Simd::splat(MAX);
let unsat = x - max;
let sat = x.saturating_sub(max);
assert_eq!(unsat, Simd::from_array([1, MAX, MIN, 0]));
assert_eq!(sat, Simd::from_array([MIN, MIN, MIN, 0]));

pub fn abs(self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Lanewise absolute value, implemented in Rust. Every lane becomes its absolute value.

Examples

let xs = Simd::from_array([MIN, MIN +1, -5, 0]);
assert_eq!(xs.abs(), Simd::from_array([MIN, MAX, 5, 0]));

pub fn saturating_abs(self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Lanewise saturating absolute value, implemented in Rust. As abs(), except the MIN value becomes MAX instead of itself.

Examples

let xs = Simd::from_array([MIN, -2, 0, 3]);
let unsat = xs.abs();
let sat = xs.saturating_abs();
assert_eq!(unsat, Simd::from_array([MIN, 2, 0, 3]));
assert_eq!(sat, Simd::from_array([MAX, 2, 0, 3]));

pub fn saturating_neg(self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Lanewise saturating negation, implemented in Rust. As neg(), except the MIN value becomes MAX instead of itself.

Examples

let x = Simd::from_array([MIN, -2, 3, MAX]);
let unsat = -x;
let sat = x.saturating_neg();
assert_eq!(unsat, Simd::from_array([MIN, 2, -3, MIN + 1]));
assert_eq!(sat, Simd::from_array([MAX, 2, -3, MIN + 1]));

impl<const LANES: usize> Simd<isize, LANES> where
LaneCount<LANES>: SupportedLaneCount,

pub fn saturating_add(self, second: Self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Lanewise saturating add.

Examples

let x = Simd::from_array([MIN, 0, 1, MAX]);
let max = Simd::splat(MAX);
let unsat = x + max;
let sat = x.saturating_add(max);
assert_eq!(unsat, Simd::from_array([-1, MAX, MIN, -2]));
assert_eq!(sat, Simd::from_array([-1, MAX, MAX, MAX]));

pub fn saturating_sub(self, second: Self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Lanewise saturating subtract.

Examples

let x = Simd::from_array([MIN, -2, -1, MAX]);
let max = Simd::splat(MAX);
let unsat = x - max;
let sat = x.saturating_sub(max);
assert_eq!(unsat, Simd::from_array([1, MAX, MIN, 0]));
assert_eq!(sat, Simd::from_array([MIN, MIN, MIN, 0]));

pub fn abs(self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Lanewise absolute value, implemented in Rust. Every lane becomes its absolute value.

Examples

let xs = Simd::from_array([MIN, MIN +1, -5, 0]);
assert_eq!(xs.abs(), Simd::from_array([MIN, MAX, 5, 0]));

pub fn saturating_abs(self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Lanewise saturating absolute value, implemented in Rust. As abs(), except the MIN value becomes MAX instead of itself.

Examples

let xs = Simd::from_array([MIN, -2, 0, 3]);
let unsat = xs.abs();
let sat = xs.saturating_abs();
assert_eq!(unsat, Simd::from_array([MIN, 2, 0, 3]));
assert_eq!(sat, Simd::from_array([MAX, 2, 0, 3]));

pub fn saturating_neg(self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Lanewise saturating negation, implemented in Rust. As neg(), except the MIN value becomes MAX instead of itself.

Examples

let x = Simd::from_array([MIN, -2, 3, MAX]);
let unsat = -x;
let sat = x.saturating_neg();
assert_eq!(unsat, Simd::from_array([MIN, 2, -3, MIN + 1]));
assert_eq!(sat, Simd::from_array([MAX, 2, -3, MIN + 1]));

impl<const LANES: usize> Simd<f32, LANES> where
LaneCount<LANES>: SupportedLaneCount,

pub unsafe fn to_int_unchecked<I>(self) -> Simd<I, LANES> where
f32: FloatToInt<I>,
I: SimdElement,

🔬 This is a nightly-only experimental API. (portable_simd)

Rounds toward zero and converts to the same-width integer type, assuming that the value is finite and fits in that type.

Safety

The value must:

Not be NaN

Not be infinite

Be representable in the return type, after truncating off its fractional part

impl<const LANES: usize> Simd<f64, LANES> where
LaneCount<LANES>: SupportedLaneCount,

pub unsafe fn to_int_unchecked<I>(self) -> Simd<I, LANES> where
f64: FloatToInt<I>,
I: SimdElement,

🔬 This is a nightly-only experimental API. (portable_simd)

Rounds toward zero and converts to the same-width integer type, assuming that the value is finite and fits in that type.

Safety

The value must:

Not be NaN

Not be infinite

Be representable in the return type, after truncating off its fractional part

impl<const LANES: usize> Simd<f32, LANES> where
LaneCount<LANES>: SupportedLaneCount,

pub fn to_bits(self) -> Simd<u32, LANES>

🔬 This is a nightly-only experimental API. (portable_simd)

Raw transmutation to an unsigned integer vector type with the same size and number of lanes.

pub fn from_bits(bits: Simd<u32, LANES>) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Raw transmutation from an unsigned integer vector type with the same size and number of lanes.

pub fn abs(self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Produces a vector where every lane has the absolute value of the equivalently-indexed lane in self.

pub fn recip(self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Takes the reciprocal (inverse) of each lane, 1/x.

pub fn to_degrees(self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Converts each lane from radians to degrees.

pub fn to_radians(self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Converts each lane from degrees to radians.

pub fn is_sign_positive(self) -> Mask<i32, LANES>

🔬 This is a nightly-only experimental API. (portable_simd)

Returns true for each lane if it has a positive sign, including +0.0, NaNs with positive sign bit and positive infinity.

pub fn is_sign_negative(self) -> Mask<i32, LANES>

🔬 This is a nightly-only experimental API. (portable_simd)

Returns true for each lane if it has a negative sign, including -0.0, NaNs with negative sign bit and negative infinity.

pub fn is_nan(self) -> Mask<i32, LANES>

🔬 This is a nightly-only experimental API. (portable_simd)

Returns true for each lane if its value is NaN.

pub fn is_infinite(self) -> Mask<i32, LANES>

🔬 This is a nightly-only experimental API. (portable_simd)

Returns true for each lane if its value is positive infinity or negative infinity.

pub fn is_finite(self) -> Mask<i32, LANES>

🔬 This is a nightly-only experimental API. (portable_simd)

Returns true for each lane if its value is neither infinite nor NaN.

pub fn is_subnormal(self) -> Mask<i32, LANES>

🔬 This is a nightly-only experimental API. (portable_simd)

Returns true for each lane if its value is subnormal.

pub fn is_normal(self) -> Mask<i32, LANES>

🔬 This is a nightly-only experimental API. (portable_simd)

Returns true for each lane if its value is neither zero, infinite, subnormal, nor NaN.

pub fn signum(self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Replaces each lane with a number that represents its sign.

1.0 if the number is positive, +0.0, or INFINITY

-1.0 if the number is negative, -0.0, or NEG_INFINITY

NAN if the number is NAN

pub fn copysign(self, sign: Self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Returns each lane with the magnitude of self and the sign of sign.

If any lane is a NAN, then a NAN with the sign of sign is returned.

pub fn min(self, other: Self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Returns the minimum of each lane.

If one of the values is NAN, then the other value is returned.

pub fn max(self, other: Self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Returns the maximum of each lane.

If one of the values is NAN, then the other value is returned.

pub fn clamp(self, min: Self, max: Self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Restrict each lane to a certain interval unless it is NaN.

For each lane in self, returns the corresponding lane in max if the lane is greater than max, and the corresponding lane in min if the lane is less than min. Otherwise returns the lane in self.

impl<const LANES: usize> Simd<f64, LANES> where
LaneCount<LANES>: SupportedLaneCount,

pub fn to_bits(self) -> Simd<u64, LANES>

🔬 This is a nightly-only experimental API. (portable_simd)

Raw transmutation to an unsigned integer vector type with the same size and number of lanes.

pub fn from_bits(bits: Simd<u64, LANES>) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Raw transmutation from an unsigned integer vector type with the same size and number of lanes.

pub fn abs(self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Produces a vector where every lane has the absolute value of the equivalently-indexed lane in self.

pub fn recip(self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Takes the reciprocal (inverse) of each lane, 1/x.

pub fn to_degrees(self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Converts each lane from radians to degrees.

pub fn to_radians(self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Converts each lane from degrees to radians.

pub fn is_sign_positive(self) -> Mask<i64, LANES>

🔬 This is a nightly-only experimental API. (portable_simd)

Returns true for each lane if it has a positive sign, including +0.0, NaNs with positive sign bit and positive infinity.

pub fn is_sign_negative(self) -> Mask<i64, LANES>

🔬 This is a nightly-only experimental API. (portable_simd)

Returns true for each lane if it has a negative sign, including -0.0, NaNs with negative sign bit and negative infinity.

pub fn is_nan(self) -> Mask<i64, LANES>

🔬 This is a nightly-only experimental API. (portable_simd)

Returns true for each lane if its value is NaN.

pub fn is_infinite(self) -> Mask<i64, LANES>

🔬 This is a nightly-only experimental API. (portable_simd)

Returns true for each lane if its value is positive infinity or negative infinity.

pub fn is_finite(self) -> Mask<i64, LANES>

🔬 This is a nightly-only experimental API. (portable_simd)

Returns true for each lane if its value is neither infinite nor NaN.

pub fn is_subnormal(self) -> Mask<i64, LANES>

🔬 This is a nightly-only experimental API. (portable_simd)

Returns true for each lane if its value is subnormal.

pub fn is_normal(self) -> Mask<i64, LANES>

🔬 This is a nightly-only experimental API. (portable_simd)

Returns true for each lane if its value is neither zero, infinite, subnormal, nor NaN.

pub fn signum(self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Replaces each lane with a number that represents its sign.

1.0 if the number is positive, +0.0, or INFINITY

-1.0 if the number is negative, -0.0, or NEG_INFINITY

NAN if the number is NAN

pub fn copysign(self, sign: Self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Returns each lane with the magnitude of self and the sign of sign.

If any lane is a NAN, then a NAN with the sign of sign is returned.

pub fn min(self, other: Self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Returns the minimum of each lane.

If one of the values is NAN, then the other value is returned.

pub fn max(self, other: Self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Returns the maximum of each lane.

If one of the values is NAN, then the other value is returned.

pub fn clamp(self, min: Self, max: Self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Restrict each lane to a certain interval unless it is NaN.

For each lane in self, returns the corresponding lane in max if the lane is greater than max, and the corresponding lane in min if the lane is less than min. Otherwise returns the lane in self.

impl<const LANES: usize> Simd<isize, LANES> where
LaneCount<LANES>: SupportedLaneCount,

pub fn is_positive(self) -> Mask<isize, LANES>

🔬 This is a nightly-only experimental API. (portable_simd)

Returns true for each positive lane and false if it is zero or negative.

pub fn is_negative(self) -> Mask<isize, LANES>

🔬 This is a nightly-only experimental API. (portable_simd)

Returns true for each negative lane and false if it is zero or positive.

pub fn signum(self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Returns numbers representing the sign of each lane.

0 if the number is zero

1 if the number is positive

-1 if the number is negative

impl<const LANES: usize> Simd<i16, LANES> where
LaneCount<LANES>: SupportedLaneCount,

pub fn is_positive(self) -> Mask<i16, LANES>

🔬 This is a nightly-only experimental API. (portable_simd)

Returns true for each positive lane and false if it is zero or negative.

pub fn is_negative(self) -> Mask<i16, LANES>

🔬 This is a nightly-only experimental API. (portable_simd)

Returns true for each negative lane and false if it is zero or positive.

pub fn signum(self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Returns numbers representing the sign of each lane.

0 if the number is zero

1 if the number is positive

-1 if the number is negative

impl<const LANES: usize> Simd<i32, LANES> where
LaneCount<LANES>: SupportedLaneCount,

pub fn is_positive(self) -> Mask<i32, LANES>

🔬 This is a nightly-only experimental API. (portable_simd)

Returns true for each positive lane and false if it is zero or negative.

pub fn is_negative(self) -> Mask<i32, LANES>

🔬 This is a nightly-only experimental API. (portable_simd)

Returns true for each negative lane and false if it is zero or positive.

pub fn signum(self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Returns numbers representing the sign of each lane.

0 if the number is zero

1 if the number is positive

-1 if the number is negative

impl<const LANES: usize> Simd<i64, LANES> where
LaneCount<LANES>: SupportedLaneCount,

pub fn is_positive(self) -> Mask<i64, LANES>

🔬 This is a nightly-only experimental API. (portable_simd)

Returns true for each positive lane and false if it is zero or negative.

pub fn is_negative(self) -> Mask<i64, LANES>

🔬 This is a nightly-only experimental API. (portable_simd)

Returns true for each negative lane and false if it is zero or positive.

pub fn signum(self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Returns numbers representing the sign of each lane.

0 if the number is zero

1 if the number is positive

-1 if the number is negative

impl<const LANES: usize> Simd<i8, LANES> where
LaneCount<LANES>: SupportedLaneCount,

pub fn is_positive(self) -> Mask<i8, LANES>

🔬 This is a nightly-only experimental API. (portable_simd)

Returns true for each positive lane and false if it is zero or negative.

pub fn is_negative(self) -> Mask<i8, LANES>

🔬 This is a nightly-only experimental API. (portable_simd)

Returns true for each negative lane and false if it is zero or positive.

pub fn signum(self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Returns numbers representing the sign of each lane.

0 if the number is zero

1 if the number is positive

-1 if the number is negative

impl<T, const LANES: usize> Simd<T, LANES> where
LaneCount<LANES>: SupportedLaneCount,
T: SimdElement,

pub const LANES: usize

🔬 This is a nightly-only experimental API. (portable_simd)

Number of lanes in this vector.

pub const fn lanes(&self) -> usize

🔬 This is a nightly-only experimental API. (portable_simd)

Get the number of lanes in this vector.

pub const fn splat(value: T) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Construct a SIMD vector by setting all lanes to the given value.

pub const fn as_array(&self) -> &[T; LANES]

🔬 This is a nightly-only experimental API. (portable_simd)

Returns an array reference containing the entire SIMD vector.

pub fn as_mut_array(&mut self) -> &mut [T; LANES]

🔬 This is a nightly-only experimental API. (portable_simd)

Returns a mutable array reference containing the entire SIMD vector.

pub const fn from_array(array: [T; LANES]) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Converts an array to a SIMD vector.

pub const fn to_array(self) -> [T; LANES]

🔬 This is a nightly-only experimental API. (portable_simd)

Converts a SIMD vector to an array.

pub const fn from_slice(slice: &[T]) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Converts a slice to a SIMD vector containing slice[..LANES]

Panics

from_slice will panic if the slice’s len is less than the vector’s Simd::LANES.

pub fn cast<U: SimdElement>(self) -> Simd<U, LANES>

🔬 This is a nightly-only experimental API. (portable_simd)

Performs lanewise conversion of a SIMD vector’s elements to another SIMD-valid type. This follows the semantics of Rust’s as conversion for casting integers to unsigned integers (interpreting as the other type, so -1 to MAX), and from floats to integers (truncating, or saturating at the limits) for each lane, or vice versa.

Examples

let floats: Simd<f32, 4> = Simd::from_array([1.9, -4.5, f32::INFINITY, f32::NAN]);
let ints = floats.cast::<i32>();
assert_eq!(ints, Simd::from_array([1, -4, i32::MAX, 0]));

// Formally equivalent, but `Simd::cast` can optimize better.
assert_eq!(ints, Simd::from_array(floats.to_array().map(|x| x as i32)));

// The float conversion does not round-trip.
let floats_again = ints.cast();
assert_ne!(floats, floats_again);
assert_eq!(floats_again, Simd::from_array([1.0, -4.0, 2147483647.0, 0.0]));

pub fn gather_or(slice: &[T], idxs: Simd<usize, LANES>, or: Self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Reads from potentially discontiguous indices in slice to construct a SIMD vector. If an index is out-of-bounds, the lane is instead selected from the or vector.

Examples

let vec: Vec<i32> = vec![10, 11, 12, 13, 14, 15, 16, 17, 18];
let idxs = Simd::from_array([9, 3, 0, 5]);
let alt = Simd::from_array([-5, -4, -3, -2]);

let result = Simd::gather_or(&vec, idxs, alt); // Note the lane that is out-of-bounds.
assert_eq!(result, Simd::from_array([-5, 13, 10, 15]));

pub fn gather_or_default(slice: &[T], idxs: Simd<usize, LANES>) -> Self where
T: Default,

🔬 This is a nightly-only experimental API. (portable_simd)

Reads from potentially discontiguous indices in slice to construct a SIMD vector. If an index is out-of-bounds, the lane is set to the default value for the type.

Examples

let vec: Vec<i32> = vec![10, 11, 12, 13, 14, 15, 16, 17, 18];
let idxs = Simd::from_array([9, 3, 0, 5]);

let result = Simd::gather_or_default(&vec, idxs); // Note the lane that is out-of-bounds.
assert_eq!(result, Simd::from_array([0, 13, 10, 15]));

pub fn gather_select(
slice: &[T],
enable: Mask<isize, LANES>,
idxs: Simd<usize, LANES>,
or: Self
) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Reads from potentially discontiguous indices in slice to construct a SIMD vector. The mask enables all true lanes and disables all false lanes. If an index is disabled or is out-of-bounds, the lane is selected from the or vector.

Examples

let vec: Vec<i32> = vec![10, 11, 12, 13, 14, 15, 16, 17, 18];
let idxs = Simd::from_array([9, 3, 0, 5]);
let alt = Simd::from_array([-5, -4, -3, -2]);
let enable = Mask::from_array([true, true, true, false]); // Note the mask of the last lane.

let result = Simd::gather_select(&vec, enable, idxs, alt); // Note the lane that is out-of-bounds.
assert_eq!(result, Simd::from_array([-5, 13, 10, -2]));

pub unsafe fn gather_select_unchecked(
slice: &[T],
enable: Mask<isize, LANES>,
idxs: Simd<usize, LANES>,
or: Self
) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Reads from potentially discontiguous indices in slice to construct a SIMD vector. The mask enables all true lanes and disables all false lanes. If an index is disabled, the lane is selected from the or vector.

Safety

Calling this function with an enabled out-of-bounds index is undefined behavior even if the resulting value is not used.

Examples

let vec: Vec<i32> = vec![10, 11, 12, 13, 14, 15, 16, 17, 18];
let idxs = Simd::from_array([9, 3, 0, 5]);
let alt = Simd::from_array([-5, -4, -3, -2]);
let enable = Mask::from_array([true, true, true, false]); // Note the final mask lane.
// If this mask was used to gather, it would be unsound. Let's fix that.
let enable = enable & idxs.lanes_lt(Simd::splat(vec.len()));

// We have masked the OOB lane, so it's safe to gather now.
let result = unsafe { Simd::gather_select_unchecked(&vec, enable, idxs, alt) };
assert_eq!(result, Simd::from_array([-5, 13, 10, -2]));

pub fn scatter(self, slice: &mut [T], idxs: Simd<usize, LANES>)

🔬 This is a nightly-only experimental API. (portable_simd)

Writes the values in a SIMD vector to potentially discontiguous indices in slice. If two lanes in the scattered vector would write to the same index only the last lane is guaranteed to actually be written.

Examples

let mut vec: Vec<i32> = vec![10, 11, 12, 13, 14, 15, 16, 17, 18];
let idxs = Simd::from_array([9, 3, 0, 0]);
let vals = Simd::from_array([-27, 82, -41, 124]);

vals.scatter(&mut vec, idxs); // index 0 receives two writes.
assert_eq!(vec, vec![124, 11, 12, 82, 14, 15, 16, 17, 18]);

pub fn scatter_select(
self,
slice: &mut [T],
enable: Mask<isize, LANES>,
idxs: Simd<usize, LANES>
)

🔬 This is a nightly-only experimental API. (portable_simd)

Writes the values in a SIMD vector to multiple potentially discontiguous indices in slice. The mask enables all true lanes and disables all false lanes. If an enabled index is out-of-bounds, the lane is not written. If two enabled lanes in the scattered vector would write to the same index, only the last lane is guaranteed to actually be written.

Examples

let mut vec: Vec<i32> = vec![10, 11, 12, 13, 14, 15, 16, 17, 18];
let idxs = Simd::from_array([9, 3, 0, 0]);
let vals = Simd::from_array([-27, 82, -41, 124]);
let enable = Mask::from_array([true, true, true, false]); // Note the mask of the last lane.

vals.scatter_select(&mut vec, enable, idxs); // index 0's second write is masked, thus omitted.
assert_eq!(vec, vec![-41, 11, 12, 82, 14, 15, 16, 17, 18]);

pub unsafe fn scatter_select_unchecked(
self,
slice: &mut [T],
enable: Mask<isize, LANES>,
idxs: Simd<usize, LANES>
)

🔬 This is a nightly-only experimental API. (portable_simd)

Writes the values in a SIMD vector to multiple potentially discontiguous indices in slice. The mask enables all true lanes and disables all false lanes. If two enabled lanes in the scattered vector would write to the same index, only the last lane is guaranteed to actually be written.

Safety

Calling this function with an enabled out-of-bounds index is undefined behavior, and may lead to memory corruption.

Examples

let mut vec: Vec<i32> = vec![10, 11, 12, 13, 14, 15, 16, 17, 18];
let idxs = Simd::from_array([9, 3, 0, 0]);
let vals = Simd::from_array([-27, 82, -41, 124]);
let enable = Mask::from_array([true, true, true, false]); // Note the mask of the last lane.
// If this mask was used to scatter, it would be unsound. Let's fix that.
let enable = enable & idxs.lanes_lt(Simd::splat(vec.len()));

// We have masked the OOB lane, so it's safe to scatter now.
unsafe { vals.scatter_select_unchecked(&mut vec, enable, idxs); }
// index 0's second write is masked, thus was omitted.
assert_eq!(vec, vec![-41, 11, 12, 82, 14, 15, 16, 17, 18]);