core_simd::simd
Trait ToBytes

pub trait ToBytes: Sealed {
    type Bytes: Copy + Unpin + Send + Sync + AsRef<[u8]> + AsMut<[u8]> + SimdUint<Scalar = u8> + 'static;

    // Required methods
    fn to_ne_bytes(self) -> Self::Bytes;
    fn to_be_bytes(self) -> Self::Bytes;
    fn to_le_bytes(self) -> Self::Bytes;
    fn from_ne_bytes(bytes: Self::Bytes) -> Self;
    fn from_be_bytes(bytes: Self::Bytes) -> Self;
    fn from_le_bytes(bytes: Self::Bytes) -> Self;
}

๐ฌThis is a nightly-only experimental API. (portable_simd)

Converts SIMD vectors to vectors of bytes

Required Associated Types

type Bytes: Copy + Unpin + Send + Sync + AsRef<[u8]> + AsMut<[u8]> + SimdUint<Scalar = u8> + 'static

๐ฌThis is a nightly-only experimental API. (portable_simd)

This type, reinterpreted as bytes.

Required Methods

fn to_ne_bytes(self) -> Self::Bytes

๐ฌThis is a nightly-only experimental API. (portable_simd)

Returns the memory representation of this integer as a byte array in native byte order.

fn to_be_bytes(self) -> Self::Bytes

๐ฌThis is a nightly-only experimental API. (portable_simd)

Returns the memory representation of this integer as a byte array in big-endian (network) byte order.

fn to_le_bytes(self) -> Self::Bytes

๐ฌThis is a nightly-only experimental API. (portable_simd)

Returns the memory representation of this integer as a byte array in little-endian byte order.

fn from_ne_bytes(bytes: Self::Bytes) -> Self

๐ฌThis is a nightly-only experimental API. (portable_simd)

Creates a native endian integer value from its memory representation as a byte array in native endianness.

fn from_be_bytes(bytes: Self::Bytes) -> Self

๐ฌThis is a nightly-only experimental API. (portable_simd)

Creates an integer value from its representation as a byte array in big endian.

fn from_le_bytes(bytes: Self::Bytes) -> Self

๐ฌThis is a nightly-only experimental API. (portable_simd)

Creates an integer value from its representation as a byte array in little endian.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.