Type Definition core_simd::simd::mask16x32

pub type mask16x32 = Mask<i16, 32>;

🔬This is a nightly-only experimental API. (portable_simd)

A SIMD mask with 32 elements for vectors with 16-bit element types.

The layout of this type is unspecified, and may change between platforms and/or Rust versions, and code should not assume that it is equivalent to [i16; 32].