Type Alias core_simd::simd::mask32x8

pub type mask32x8 = Mask<i32, 8>;

ðŸ”¬This is a nightly-only experimental API. (portable_simd)

A SIMD mask with eight elements for vectors with 32-bit element types.

The layout of this type is unspecified, and may change between platforms and/or Rust versions, and code should not assume that it is equivalent to [i32; 8].

Aliased Type

struct mask32x8(/* private fields */);

Implementations

impl<T, const LANES: usize> Mask<T, LANES>where T: MaskElement, LaneCount<LANES>: SupportedLaneCount,

pub fn splat(value: bool) -> Self

ðŸ”¬This is a nightly-only experimental API. (portable_simd)

Construct a mask by setting all lanes to the given value.

pub fn from_array(array: [bool; LANES]) -> Self

ðŸ”¬This is a nightly-only experimental API. (portable_simd)

Converts an array of bools to a SIMD mask.

pub fn to_array(self) -> [bool; LANES]

ðŸ”¬This is a nightly-only experimental API. (portable_simd)

Converts a SIMD mask to an array of bools.

pub unsafe fn from_int_unchecked(value: Simd<T, LANES>) -> Self

ðŸ”¬This is a nightly-only experimental API. (portable_simd)

Converts a vector of integers to a mask, where 0 represents false and -1 represents true.

Safety

All lanes must be either 0 or -1.

pub fn from_int(value: Simd<T, LANES>) -> Self

ðŸ”¬This is a nightly-only experimental API. (portable_simd)

Converts a vector of integers to a mask, where 0 represents false and -1 represents true.

Panics

Panics if any lane is not 0 or -1.

pub fn to_int(self) -> Simd<T, LANES>

ðŸ”¬This is a nightly-only experimental API. (portable_simd)

Converts the mask to a vector of integers, where 0 represents false and -1 represents true.

pub fn cast<U: MaskElement>(self) -> Mask<U, LANES>

ðŸ”¬This is a nightly-only experimental API. (portable_simd)

Converts the mask to a mask of any other lane size.

pub unsafe fn test_unchecked(&self, lane: usize) -> bool

ðŸ”¬This is a nightly-only experimental API. (portable_simd)

Tests the value of the specified lane.

Safety

lane must be less than LANES.

pub fn test(&self, lane: usize) -> bool

ðŸ”¬This is a nightly-only experimental API. (portable_simd)

Tests the value of the specified lane.

Panics

Panics if lane is greater than or equal to the number of lanes in the vector.

pub unsafe fn set_unchecked(&mut self, lane: usize, value: bool)

ðŸ”¬This is a nightly-only experimental API. (portable_simd)

Sets the value of the specified lane.

Safety

lane must be less than LANES.

pub fn set(&mut self, lane: usize, value: bool)

ðŸ”¬This is a nightly-only experimental API. (portable_simd)

Sets the value of the specified lane.

Panics

Panics if lane is greater than or equal to the number of lanes in the vector.

pub fn any(self) -> bool

ðŸ”¬This is a nightly-only experimental API. (portable_simd)

Returns true if any lane is set, or false otherwise.

pub fn all(self) -> bool

ðŸ”¬This is a nightly-only experimental API. (portable_simd)

Returns true if all lanes are set, or false otherwise.

impl<T, const LANES: usize> Mask<T, LANES>where T: MaskElement, LaneCount<LANES>: SupportedLaneCount,

pub fn select<U>( self, true_values: Simd<U, LANES>, false_values: Simd<U, LANES> ) -> Simd<U, LANES>where U: SimdElement<Mask = T>,

ðŸ”¬This is a nightly-only experimental API. (portable_simd)

Choose lanes from two vectors.

For each lane in the mask, choose the corresponding lane from true_values if that lane mask is true, and false_values if that lane mask is false.

Examples

let a = Simd::from_array([0, 1, 2, 3]);
let b = Simd::from_array([4, 5, 6, 7]);
let mask = Mask::from_array([true, false, false, true]);
let c = mask.select(a, b);
assert_eq!(c.to_array(), [0, 5, 6, 3]);

pub fn select_mask(self, true_values: Self, false_values: Self) -> Self

ðŸ”¬This is a nightly-only experimental API. (portable_simd)

Choose lanes from two masks.

For each lane in the mask, choose the corresponding lane from true_values if that lane mask is true, and false_values if that lane mask is false.

Examples

let a = Mask::<i32, 4>::from_array([true, true, false, false]);
let b = Mask::<i32, 4>::from_array([false, false, true, true]);
let mask = Mask::<i32, 4>::from_array([true, false, false, true]);
let c = mask.select_mask(a, b);
assert_eq!(c.to_array(), [true, false, true, false]);