Type Definition core_simd::simd::mask8x32

pub type mask8x32 = Mask<i8, 32>;

🔬 This is a nightly-only experimental API. (portable_simd)

A mask for SIMD vectors with 32 elements of 8 bits.

The layout of this type is unspecified, and may change between platforms and/or Rust versions, and code should not assume that it is equivalent to [i8; 32].