Struct core_simd::Mask16

pub struct Mask16<const LANES: usize>(_)
 where
    SimdI16<LANES>: LanesAtMost64;

Mask for vectors with LANES 16-bit elements.

The layout of this type is unspecified.

Implementations

impl<const LANES: usize> Mask16<LANES> where
SimdI16<LANES>: LanesAtMost64,

pub fn splat(value: bool) -> Self

Construct a mask by setting all lanes to the given value.

pub fn test(&self, lane: usize) -> bool

Tests the value of the specified lane.

Panics

Panics if lane is greater than or equal to the number of lanes in the vector.

pub fn set(&mut self, lane: usize, value: bool)

Sets the value of the specified lane.

Panics

Panics if lane is greater than or equal to the number of lanes in the vector.