Struct core_simd::Mask32

pub struct Mask32<const LANES: usize>(_)
 where
    SimdI32<LANES>: LanesAtMost32,
    Self: Mask;

Mask for vectors with LANES 32-bit elements.

The layout of this type is unspecified.

Implementations

impl<const LANES: usize> Mask32<LANES> where
Self: Mask,
SimdI32<LANES>: LanesAtMost32,

pub fn select<S: Select<Self>>(self, true_values: S, false_values: S) -> S

Choose lanes from two vectors.

For each lane in the mask, choose the corresponding lane from true_values if that lane mask is true, and false_values if that lane mask is false.

let a = SimdI32::from_array([0, 1, 2, 3]);
let b = SimdI32::from_array([4, 5, 6, 7]);
let mask = Mask32::from_array([true, false, false, true]);
let c = mask.select(a, b);
assert_eq!(c.to_array(), [0, 5, 6, 3]);

select can also be used on masks:

let a = Mask32::from_array([true, true, false, false]);
let b = Mask32::from_array([false, false, true, true]);
let mask = Mask32::from_array([true, false, false, true]);
let c = mask.select(a, b);
assert_eq!(c.to_array(), [true, false, true, false]);

impl<const LANES: usize> Mask32<LANES> where
SimdI32<LANES>: LanesAtMost32,
Mask32<LANES>: Mask,

pub fn any(self) -> bool

Returns true if any lane is set, or false otherwise.

pub fn all(self) -> bool

Returns true if all lanes are set, or false otherwise.

impl<const LANES: usize> Mask32<LANES> where
SimdI32<LANES>: LanesAtMost32,
Self: Mask,

pub fn splat(value: bool) -> Self

Construct a mask by setting all lanes to the given value.

pub fn from_array(array: [bool; LANES]) -> Self

Converts an array to a SIMD vector.

pub fn to_array(self) -> [bool; LANES]

Converts a SIMD vector to an array.

pub unsafe fn from_int_unchecked(value: SimdI32<LANES>) -> Self

Converts a vector of integers to a mask, where 0 represents false and -1 represents true.

Safety

All lanes must be either 0 or -1.

pub fn from_int(value: SimdI32<LANES>) -> Self

Converts a vector of integers to a mask, where 0 represents false and -1 represents true.

Panics

Panics if any lane is not 0 or -1.

pub fn to_int(self) -> SimdI32<LANES>

Converts the mask to a vector of integers, where 0 represents false and -1 represents true.

pub unsafe fn test_unchecked(&self, lane: usize) -> bool

Tests the value of the specified lane.

Safety

lane must be less than LANES.

pub fn test(&self, lane: usize) -> bool

Tests the value of the specified lane.

Panics

Panics if lane is greater than or equal to the number of lanes in the vector.

pub unsafe fn set_unchecked(&mut self, lane: usize, value: bool)

Sets the value of the specified lane.

Safety

lane must be less than LANES.

pub fn set(&mut self, lane: usize, value: bool)

Sets the value of the specified lane.

Panics

Panics if lane is greater than or equal to the number of lanes in the vector.

pub fn to_bitmask(self) -> <Self as Mask>::BitMask

Convert this mask to a bitmask, with one bit set per lane.

pub fn from_bitmask(bitmask: <Self as Mask>::BitMask) -> Self

Convert a bitmask to a mask.