Struct core_simd::Mask64

pub struct Mask64<const LANES: usize>(_)
 where
    SimdI64<LANES>: LanesAtMost32;

Mask for vectors with LANES 64-bit elements.

The layout of this type is unspecified.

Implementations

impl<const LANES: usize> Mask64<LANES> where
SimdI64<LANES>: LanesAtMost32,

pub fn splat(value: bool) -> Self

Construct a mask by setting all lanes to the given value.

pub fn from_array(array: [bool; LANES]) -> Self

Converts an array to a SIMD vector.

pub fn to_array(self) -> [bool; LANES]

Converts a SIMD vector to an array.

pub fn test(&self, lane: usize) -> bool

Tests the value of the specified lane.

Panics

Panics if lane is greater than or equal to the number of lanes in the vector.

pub fn set(&mut self, lane: usize, value: bool)

Sets the value of the specified lane.

Panics

Panics if lane is greater than or equal to the number of lanes in the vector.