Struct core_simd::SimdF32

#[repr(simd)]pub struct SimdF32<const LANES: usize>(_)
 where
    Self: LanesAtMost64;

A SIMD vector of containing LANES f32 values.

Implementations

impl<const LANES: usize> SimdF32<LANES> where
Self: LanesAtMost64,

#[must_use = "method returns a new vector and does not mutate the original value"]pub fn floor(self) -> Self

Returns the largest integer less than or equal to each lane.

#[must_use = "method returns a new vector and does not mutate the original value"]pub fn ceil(self) -> Self

Returns the smallest integer greater than or equal to each lane.

impl<const LANES: usize> SimdF32<LANES> where
Self: LanesAtMost64,
SimdI32<LANES>: LanesAtMost64,

pub unsafe fn to_int_unchecked(self) -> SimdI32<LANES>

Rounds toward zero and converts to the same-width integer type, assuming that the value is finite and fits in that type.

Safety

The value must:

Not be NaN

Not be infinite

Be representable in the return type, after truncating off its fractional part

pub fn round_from_int(value: SimdI32<LANES>) -> Self

Creates a floating-point vector from an integer vector. Rounds values that are not exactly representable.

impl<const LANES: usize> SimdF32<LANES> where
SimdF32<LANES>: LanesAtMost64,
SimdI32<LANES>: LanesAtMost64,

pub fn lanes_eq(&self, other: &Self) -> Mask32<LANES>

Test if each lane is equal to the corresponding lane in other.

pub fn lanes_ne(&self, other: &Self) -> Mask32<LANES>

Test if each lane is not equal to the corresponding lane in other.

pub fn lanes_lt(&self, other: &Self) -> Mask32<LANES>

Test if each lane is less than the corresponding lane in other.

pub fn lanes_gt(&self, other: &Self) -> Mask32<LANES>

Test if each lane is greater than the corresponding lane in other.

pub fn lanes_le(&self, other: &Self) -> Mask32<LANES>

Test if each lane is less than or equal to the corresponding lane in other.

pub fn lanes_ge(&self, other: &Self) -> Mask32<LANES>

Test if each lane is greater than or equal to the corresponding lane in other.

impl<const LANES: usize> SimdF32<LANES> where
Self: LanesAtMost64,

pub const fn splat(value: f32) -> Self

Construct a SIMD vector by setting all lanes to the given value.

pub const fn as_slice(&self) -> &[f32]

Returns a slice containing the entire SIMD vector.

pub fn as_mut_slice(&mut self) -> &mut [f32]

Returns a mutable slice containing the entire SIMD vector.

pub const fn from_array(array: [f32; LANES]) -> Self

Converts an array to a SIMD vector.

pub const fn to_array(self) -> [f32; LANES]

Converts a SIMD vector to an array.

impl SimdF32<2>

pub fn shuffle<const IDX: [u32; 2]>(self, second: Self) -> Self

A const SIMD shuffle that takes 2 SIMD vectors and produces another vector, using the indices in the const parameter. The first or “self” vector will have its lanes indexed from 0, and the second vector will have its first lane indexed at $n. Indices must be in-bounds of either vector at compile time.

Some SIMD shuffle instructions can be quite slow, so avoiding them by loading data into the desired patterns in advance is preferred, but shuffles are still faster than storing and reloading from memory.

impl SimdF32<4>

pub fn shuffle<const IDX: [u32; 4]>(self, second: Self) -> Self

A const SIMD shuffle that takes 2 SIMD vectors and produces another vector, using the indices in the const parameter. The first or “self” vector will have its lanes indexed from 0, and the second vector will have its first lane indexed at $n. Indices must be in-bounds of either vector at compile time.

Some SIMD shuffle instructions can be quite slow, so avoiding them by loading data into the desired patterns in advance is preferred, but shuffles are still faster than storing and reloading from memory.

impl SimdF32<8>

pub fn shuffle<const IDX: [u32; 8]>(self, second: Self) -> Self

A const SIMD shuffle that takes 2 SIMD vectors and produces another vector, using the indices in the const parameter. The first or “self” vector will have its lanes indexed from 0, and the second vector will have its first lane indexed at $n. Indices must be in-bounds of either vector at compile time.

Some SIMD shuffle instructions can be quite slow, so avoiding them by loading data into the desired patterns in advance is preferred, but shuffles are still faster than storing and reloading from memory.

impl SimdF32<16>

pub fn shuffle<const IDX: [u32; 16]>(self, second: Self) -> Self

A const SIMD shuffle that takes 2 SIMD vectors and produces another vector, using the indices in the const parameter. The first or “self” vector will have its lanes indexed from 0, and the second vector will have its first lane indexed at $n. Indices must be in-bounds of either vector at compile time.

Some SIMD shuffle instructions can be quite slow, so avoiding them by loading data into the desired patterns in advance is preferred, but shuffles are still faster than storing and reloading from memory.

impl SimdF32<32>

pub fn shuffle<const IDX: [u32; 32]>(self, second: Self) -> Self

A const SIMD shuffle that takes 2 SIMD vectors and produces another vector, using the indices in the const parameter. The first or “self” vector will have its lanes indexed from 0, and the second vector will have its first lane indexed at $n. Indices must be in-bounds of either vector at compile time.

Some SIMD shuffle instructions can be quite slow, so avoiding them by loading data into the desired patterns in advance is preferred, but shuffles are still faster than storing and reloading from memory.

impl SimdF32<64>

pub fn shuffle<const IDX: [u32; 64]>(self, second: Self) -> Self

A const SIMD shuffle that takes 2 SIMD vectors and produces another vector, using the indices in the const parameter. The first or “self” vector will have its lanes indexed from 0, and the second vector will have its first lane indexed at $n. Indices must be in-bounds of either vector at compile time.

Some SIMD shuffle instructions can be quite slow, so avoiding them by loading data into the desired patterns in advance is preferred, but shuffles are still faster than storing and reloading from memory.

impl<const LANES: usize> SimdF32<LANES> where
Self: LanesAtMost64,
SimdU32<LANES>: LanesAtMost64,

pub fn to_bits(self) -> SimdU32<LANES>

Raw transmutation to an unsigned integer vector type with the same size and number of lanes.

pub fn from_bits(bits: SimdU32<LANES>) -> Self

Raw transmutation from an unsigned integer vector type with the same size and number of lanes.

pub fn abs(self) -> Self

Produces a vector where every lane has the absolute value of the equivalently-indexed lane in self.