Struct core_simd::SimdF64

#[repr(simd)]
pub struct SimdF64<const LANES: usize>(_)
 where
    Self: LanesAtMost32;

🔬 This is a nightly-only experimental API. (portable_simd)

A SIMD vector of containing LANES f64 values.

Implementations

impl<const LANES: usize> SimdF64<LANES> where
SimdF64<LANES>: LanesAtMost32,
SimdI64<LANES>: LanesAtMost32,
Mask64<LANES>: Mask,

pub fn lanes_eq(self, other: Self) -> Mask64<LANES>

🔬 This is a nightly-only experimental API. (portable_simd)

Test if each lane is equal to the corresponding lane in other.

pub fn lanes_ne(self, other: Self) -> Mask64<LANES>

🔬 This is a nightly-only experimental API. (portable_simd)

Test if each lane is not equal to the corresponding lane in other.

pub fn lanes_lt(self, other: Self) -> Mask64<LANES>

🔬 This is a nightly-only experimental API. (portable_simd)

Test if each lane is less than the corresponding lane in other.

pub fn lanes_gt(self, other: Self) -> Mask64<LANES>

🔬 This is a nightly-only experimental API. (portable_simd)

Test if each lane is greater than the corresponding lane in other.

pub fn lanes_le(self, other: Self) -> Mask64<LANES>

🔬 This is a nightly-only experimental API. (portable_simd)

Test if each lane is less than or equal to the corresponding lane in other.

pub fn lanes_ge(self, other: Self) -> Mask64<LANES>

🔬 This is a nightly-only experimental API. (portable_simd)

Test if each lane is greater than or equal to the corresponding lane in other.

impl<const LANES: usize> SimdF64<LANES> where
Self: LanesAtMost32,

pub fn ceil(self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Returns the smallest integer greater than or equal to each lane.

pub fn floor(self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Returns the largest integer value less than or equal to each lane.

pub fn round(self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Rounds to the nearest integer value. Ties round toward zero.

pub fn trunc(self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Returns the floating point’s integer value, with its fractional part removed.

pub fn fract(self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Returns the floating point’s fractional value, with its integer part removed.

impl<const LANES: usize> SimdF64<LANES> where
Self: LanesAtMost32,
SimdI64<LANES>: LanesAtMost32,

pub unsafe fn to_int_unchecked(self) -> SimdI64<LANES>

🔬 This is a nightly-only experimental API. (portable_simd)

Rounds toward zero and converts to the same-width integer type, assuming that the value is finite and fits in that type.

Safety

The value must:

Not be NaN

Not be infinite

Be representable in the return type, after truncating off its fractional part

pub fn round_from_int(value: SimdI64<LANES>) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Creates a floating-point vector from an integer vector. Rounds values that are not exactly representable.

impl<const LANES: usize> SimdF64<LANES> where
Self: LanesAtMost32,

pub const fn splat(value: f64) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Construct a SIMD vector by setting all lanes to the given value.

pub const fn as_slice(&self) -> &[f64]

🔬 This is a nightly-only experimental API. (portable_simd)

Returns a slice containing the entire SIMD vector.

pub fn as_mut_slice(&mut self) -> &mut [f64]

🔬 This is a nightly-only experimental API. (portable_simd)

Returns a mutable slice containing the entire SIMD vector.

pub const fn from_array(array: [f64; LANES]) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Converts an array to a SIMD vector.

pub const fn to_array(self) -> [f64; LANES]

🔬 This is a nightly-only experimental API. (portable_simd)

Converts a SIMD vector to an array.

impl SimdF64<2>

pub fn shuffle<const IDX: [u32; 2]>(self, second: Self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

A const SIMD shuffle that takes 2 SIMD vectors and produces another vector, using the indices in the const parameter. The first or “self” vector will have its lanes indexed from 0, and the second vector will have its first lane indexed at $n. Indices must be in-bounds of either vector at compile time.

Some SIMD shuffle instructions can be quite slow, so avoiding them by loading data into the desired patterns in advance is preferred, but shuffles are still faster than storing and reloading from memory.

#![feature(portable_simd)]
let a = f32x4::from_array([1.0, 2.0, 3.0, 4.0]);
let b = f32x4::from_array([5.0, 6.0, 7.0, 8.0]);
const IDXS: [u32; 4] = [4,0,3,7];
let c = f32x4::shuffle::<IDXS>(a,b);
assert_eq!(f32x4::from_array([5.0, 1.0, 4.0, 8.0]), c);

pub fn reverse(self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Reverse the order of the lanes in the vector.

pub fn interleave(self, other: Self) -> (Self, Self)

🔬 This is a nightly-only experimental API. (portable_simd)

Interleave two vectors.

Produces two vectors with lanes taken alternately from self and other.

The first result contains the first LANES / 2 lanes from self and other, alternating, starting with the first lane of self.

The second result contains the last LANES / 2 lanes from self and other, alternating, starting with the lane LANES / 2 from the start of self.

This particular permutation is efficient on many architectures.

#![feature(portable_simd)]
let a = SimdU32::from_array([0, 1, 2, 3]);
let b = SimdU32::from_array([4, 5, 6, 7]);
let (x, y) = a.interleave(b);
assert_eq!(x.to_array(), [0, 4, 1, 5]);
assert_eq!(y.to_array(), [2, 6, 3, 7]);

pub fn deinterleave(self, other: Self) -> (Self, Self)

🔬 This is a nightly-only experimental API. (portable_simd)

Deinterleave two vectors.

The first result takes every other lane of self and then other, starting with the first lane.

The second result takes every other lane of self and then other, starting with the second lane.

This particular permutation is efficient on many architectures.

#![feature(portable_simd)]
let a = SimdU32::from_array([0, 4, 1, 5]);
let b = SimdU32::from_array([2, 6, 3, 7]);
let (x, y) = a.deinterleave(b);
assert_eq!(x.to_array(), [0, 1, 2, 3]);
assert_eq!(y.to_array(), [4, 5, 6, 7]);

impl SimdF64<4>

pub fn shuffle<const IDX: [u32; 4]>(self, second: Self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

A const SIMD shuffle that takes 2 SIMD vectors and produces another vector, using the indices in the const parameter. The first or “self” vector will have its lanes indexed from 0, and the second vector will have its first lane indexed at $n. Indices must be in-bounds of either vector at compile time.

Some SIMD shuffle instructions can be quite slow, so avoiding them by loading data into the desired patterns in advance is preferred, but shuffles are still faster than storing and reloading from memory.

#![feature(portable_simd)]
let a = f32x4::from_array([1.0, 2.0, 3.0, 4.0]);
let b = f32x4::from_array([5.0, 6.0, 7.0, 8.0]);
const IDXS: [u32; 4] = [4,0,3,7];
let c = f32x4::shuffle::<IDXS>(a,b);
assert_eq!(f32x4::from_array([5.0, 1.0, 4.0, 8.0]), c);

pub fn reverse(self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Reverse the order of the lanes in the vector.

pub fn interleave(self, other: Self) -> (Self, Self)

🔬 This is a nightly-only experimental API. (portable_simd)

Interleave two vectors.

Produces two vectors with lanes taken alternately from self and other.

The first result contains the first LANES / 2 lanes from self and other, alternating, starting with the first lane of self.

The second result contains the last LANES / 2 lanes from self and other, alternating, starting with the lane LANES / 2 from the start of self.

This particular permutation is efficient on many architectures.

#![feature(portable_simd)]
let a = SimdU32::from_array([0, 1, 2, 3]);
let b = SimdU32::from_array([4, 5, 6, 7]);
let (x, y) = a.interleave(b);
assert_eq!(x.to_array(), [0, 4, 1, 5]);
assert_eq!(y.to_array(), [2, 6, 3, 7]);

pub fn deinterleave(self, other: Self) -> (Self, Self)

🔬 This is a nightly-only experimental API. (portable_simd)

Deinterleave two vectors.

The first result takes every other lane of self and then other, starting with the first lane.

The second result takes every other lane of self and then other, starting with the second lane.

This particular permutation is efficient on many architectures.

#![feature(portable_simd)]
let a = SimdU32::from_array([0, 4, 1, 5]);
let b = SimdU32::from_array([2, 6, 3, 7]);
let (x, y) = a.deinterleave(b);
assert_eq!(x.to_array(), [0, 1, 2, 3]);
assert_eq!(y.to_array(), [4, 5, 6, 7]);

impl SimdF64<8>

pub fn shuffle<const IDX: [u32; 8]>(self, second: Self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

A const SIMD shuffle that takes 2 SIMD vectors and produces another vector, using the indices in the const parameter. The first or “self” vector will have its lanes indexed from 0, and the second vector will have its first lane indexed at $n. Indices must be in-bounds of either vector at compile time.

Some SIMD shuffle instructions can be quite slow, so avoiding them by loading data into the desired patterns in advance is preferred, but shuffles are still faster than storing and reloading from memory.

#![feature(portable_simd)]
let a = f32x4::from_array([1.0, 2.0, 3.0, 4.0]);
let b = f32x4::from_array([5.0, 6.0, 7.0, 8.0]);
const IDXS: [u32; 4] = [4,0,3,7];
let c = f32x4::shuffle::<IDXS>(a,b);
assert_eq!(f32x4::from_array([5.0, 1.0, 4.0, 8.0]), c);

pub fn reverse(self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Reverse the order of the lanes in the vector.

pub fn interleave(self, other: Self) -> (Self, Self)

🔬 This is a nightly-only experimental API. (portable_simd)

Interleave two vectors.

Produces two vectors with lanes taken alternately from self and other.

The first result contains the first LANES / 2 lanes from self and other, alternating, starting with the first lane of self.

The second result contains the last LANES / 2 lanes from self and other, alternating, starting with the lane LANES / 2 from the start of self.

This particular permutation is efficient on many architectures.

#![feature(portable_simd)]
let a = SimdU32::from_array([0, 1, 2, 3]);
let b = SimdU32::from_array([4, 5, 6, 7]);
let (x, y) = a.interleave(b);
assert_eq!(x.to_array(), [0, 4, 1, 5]);
assert_eq!(y.to_array(), [2, 6, 3, 7]);

pub fn deinterleave(self, other: Self) -> (Self, Self)

🔬 This is a nightly-only experimental API. (portable_simd)

Deinterleave two vectors.

The first result takes every other lane of self and then other, starting with the first lane.

The second result takes every other lane of self and then other, starting with the second lane.

This particular permutation is efficient on many architectures.

#![feature(portable_simd)]
let a = SimdU32::from_array([0, 4, 1, 5]);
let b = SimdU32::from_array([2, 6, 3, 7]);
let (x, y) = a.deinterleave(b);
assert_eq!(x.to_array(), [0, 1, 2, 3]);
assert_eq!(y.to_array(), [4, 5, 6, 7]);

impl SimdF64<16>

pub fn shuffle<const IDX: [u32; 16]>(self, second: Self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

A const SIMD shuffle that takes 2 SIMD vectors and produces another vector, using the indices in the const parameter. The first or “self” vector will have its lanes indexed from 0, and the second vector will have its first lane indexed at $n. Indices must be in-bounds of either vector at compile time.

Some SIMD shuffle instructions can be quite slow, so avoiding them by loading data into the desired patterns in advance is preferred, but shuffles are still faster than storing and reloading from memory.

#![feature(portable_simd)]
let a = f32x4::from_array([1.0, 2.0, 3.0, 4.0]);
let b = f32x4::from_array([5.0, 6.0, 7.0, 8.0]);
const IDXS: [u32; 4] = [4,0,3,7];
let c = f32x4::shuffle::<IDXS>(a,b);
assert_eq!(f32x4::from_array([5.0, 1.0, 4.0, 8.0]), c);

pub fn reverse(self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Reverse the order of the lanes in the vector.

pub fn interleave(self, other: Self) -> (Self, Self)

🔬 This is a nightly-only experimental API. (portable_simd)

Interleave two vectors.

Produces two vectors with lanes taken alternately from self and other.

The first result contains the first LANES / 2 lanes from self and other, alternating, starting with the first lane of self.

The second result contains the last LANES / 2 lanes from self and other, alternating, starting with the lane LANES / 2 from the start of self.

This particular permutation is efficient on many architectures.

#![feature(portable_simd)]
let a = SimdU32::from_array([0, 1, 2, 3]);
let b = SimdU32::from_array([4, 5, 6, 7]);
let (x, y) = a.interleave(b);
assert_eq!(x.to_array(), [0, 4, 1, 5]);
assert_eq!(y.to_array(), [2, 6, 3, 7]);

pub fn deinterleave(self, other: Self) -> (Self, Self)

🔬 This is a nightly-only experimental API. (portable_simd)

Deinterleave two vectors.

The first result takes every other lane of self and then other, starting with the first lane.

The second result takes every other lane of self and then other, starting with the second lane.

This particular permutation is efficient on many architectures.

#![feature(portable_simd)]
let a = SimdU32::from_array([0, 4, 1, 5]);
let b = SimdU32::from_array([2, 6, 3, 7]);
let (x, y) = a.deinterleave(b);
assert_eq!(x.to_array(), [0, 1, 2, 3]);
assert_eq!(y.to_array(), [4, 5, 6, 7]);

impl SimdF64<32>

pub fn shuffle<const IDX: [u32; 32]>(self, second: Self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

A const SIMD shuffle that takes 2 SIMD vectors and produces another vector, using the indices in the const parameter. The first or “self” vector will have its lanes indexed from 0, and the second vector will have its first lane indexed at $n. Indices must be in-bounds of either vector at compile time.

Some SIMD shuffle instructions can be quite slow, so avoiding them by loading data into the desired patterns in advance is preferred, but shuffles are still faster than storing and reloading from memory.

#![feature(portable_simd)]
let a = f32x4::from_array([1.0, 2.0, 3.0, 4.0]);
let b = f32x4::from_array([5.0, 6.0, 7.0, 8.0]);
const IDXS: [u32; 4] = [4,0,3,7];
let c = f32x4::shuffle::<IDXS>(a,b);
assert_eq!(f32x4::from_array([5.0, 1.0, 4.0, 8.0]), c);

pub fn reverse(self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Reverse the order of the lanes in the vector.

pub fn interleave(self, other: Self) -> (Self, Self)

🔬 This is a nightly-only experimental API. (portable_simd)

Interleave two vectors.

Produces two vectors with lanes taken alternately from self and other.

The first result contains the first LANES / 2 lanes from self and other, alternating, starting with the first lane of self.

The second result contains the last LANES / 2 lanes from self and other, alternating, starting with the lane LANES / 2 from the start of self.

This particular permutation is efficient on many architectures.

#![feature(portable_simd)]
let a = SimdU32::from_array([0, 1, 2, 3]);
let b = SimdU32::from_array([4, 5, 6, 7]);
let (x, y) = a.interleave(b);
assert_eq!(x.to_array(), [0, 4, 1, 5]);
assert_eq!(y.to_array(), [2, 6, 3, 7]);

pub fn deinterleave(self, other: Self) -> (Self, Self)

🔬 This is a nightly-only experimental API. (portable_simd)

Deinterleave two vectors.

The first result takes every other lane of self and then other, starting with the first lane.

The second result takes every other lane of self and then other, starting with the second lane.

This particular permutation is efficient on many architectures.

#![feature(portable_simd)]
let a = SimdU32::from_array([0, 4, 1, 5]);
let b = SimdU32::from_array([2, 6, 3, 7]);
let (x, y) = a.deinterleave(b);
assert_eq!(x.to_array(), [0, 1, 2, 3]);
assert_eq!(y.to_array(), [4, 5, 6, 7]);

impl<const LANES: usize> SimdF64<LANES> where
Self: LanesAtMost32,

pub fn horizontal_sum(self) -> f64

🔬 This is a nightly-only experimental API. (portable_simd)

Horizontal add. Returns the sum of the lanes of the vector.

pub fn horizontal_product(self) -> f64

🔬 This is a nightly-only experimental API. (portable_simd)

Horizontal multiply. Returns the product of the lanes of the vector.

pub fn horizontal_max(self) -> f64

🔬 This is a nightly-only experimental API. (portable_simd)

Horizontal maximum. Returns the maximum lane in the vector.

Returns values based on equality, so a vector containing both 0. and -0. may return either. This function will not return NaN unless all lanes are NaN.

pub fn horizontal_min(self) -> f64

🔬 This is a nightly-only experimental API. (portable_simd)

Horizontal minimum. Returns the minimum lane in the vector.

Returns values based on equality, so a vector containing both 0. and -0. may return either. This function will not return NaN unless all lanes are NaN.

impl<const LANES: usize> SimdF64<LANES> where
Self: LanesAtMost32,
SimdU64<LANES>: LanesAtMost32,

pub fn to_bits(self) -> SimdU64<LANES>

🔬 This is a nightly-only experimental API. (portable_simd)

Raw transmutation to an unsigned integer vector type with the same size and number of lanes.

pub fn from_bits(bits: SimdU64<LANES>) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Raw transmutation from an unsigned integer vector type with the same size and number of lanes.

pub fn abs(self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Produces a vector where every lane has the absolute value of the equivalently-indexed lane in self.

pub fn mul_add(self, a: Self, b: Self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Fused multiply-add. Computes (self * a) + b with only one rounding error, yielding a more accurate result than an unfused multiply-add.

Using mul_add may be more performant than an unfused multiply-add if the target architecture has a dedicated fma CPU instruction. However, this is not always true, and will be heavily dependent on designing algorithms with specific target hardware in mind.

pub fn sqrt(self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Produces a vector where every lane has the square root value of the equivalently-indexed lane in self

pub fn recip(self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Takes the reciprocal (inverse) of each lane, 1/x.

pub fn to_degrees(self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Converts each lane from radians to degrees.

pub fn to_radians(self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Converts each lane from degrees to radians.

impl<const LANES: usize> SimdF64<LANES> where
Self: LanesAtMost32,
SimdU64<LANES>: LanesAtMost32,
SimdI64<LANES>: LanesAtMost32,
Mask64<LANES>: Mask,

pub fn is_sign_positive(self) -> Mask64<LANES>

🔬 This is a nightly-only experimental API. (portable_simd)

Returns true for each lane if it has a positive sign, including +0.0, NaNs with positive sign bit and positive infinity.

pub fn is_sign_negative(self) -> Mask64<LANES>

🔬 This is a nightly-only experimental API. (portable_simd)

Returns true for each lane if it has a negative sign, including -0.0, NaNs with negative sign bit and negative infinity.

pub fn is_nan(self) -> Mask64<LANES>

🔬 This is a nightly-only experimental API. (portable_simd)

Returns true for each lane if its value is NaN.

pub fn is_infinite(self) -> Mask64<LANES>

🔬 This is a nightly-only experimental API. (portable_simd)

Returns true for each lane if its value is positive infinity or negative infinity.

pub fn is_finite(self) -> Mask64<LANES>

🔬 This is a nightly-only experimental API. (portable_simd)

Returns true for each lane if its value is neither infinite nor NaN.

pub fn is_subnormal(self) -> Mask64<LANES>

🔬 This is a nightly-only experimental API. (portable_simd)

Returns true for each lane if its value is subnormal.

pub fn is_normal(self) -> Mask64<LANES>

🔬 This is a nightly-only experimental API. (portable_simd)

Returns true for each lane if its value is neither neither zero, infinite, subnormal, or NaN.

pub fn signum(self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Replaces each lane with a number that represents its sign.

1.0 if the number is positive, +0.0, or INFINITY

-1.0 if the number is negative, -0.0, or NEG_INFINITY

NAN if the number is NAN

pub fn copysign(self, sign: Self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Returns each lane with the magnitude of self and the sign of sign.

If any lane is a NAN, then a NAN with the sign of sign is returned.

pub fn min(self, other: Self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Returns the minimum of each lane.

If one of the values is NAN, then the other value is returned.

pub fn max(self, other: Self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Returns the maximum of each lane.

If one of the values is NAN, then the other value is returned.

pub fn clamp(self, min: Self, max: Self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Restrict each lane to a certain interval unless it is NaN.

For each lane in self, returns the corresponding lane in max if the lane is greater than max, and the corresponding lane in min if the lane is less than min. Otherwise returns the lane in self.