Struct core_simd::SimdI16

#[repr(simd)]
pub struct SimdI16<const LANES: usize>(_)
 where
    Self: LanesAtMost32;

A SIMD vector of containing LANES i16 values.

Implementations

impl<const LANES: usize> SimdI16<LANES> where
SimdI16<LANES>: LanesAtMost32,
SimdI16<LANES>: LanesAtMost32,

pub fn lanes_eq(self, other: Self) -> Mask16<LANES>

Test if each lane is equal to the corresponding lane in other.

pub fn lanes_ne(self, other: Self) -> Mask16<LANES>

Test if each lane is not equal to the corresponding lane in other.

pub fn lanes_lt(self, other: Self) -> Mask16<LANES>

Test if each lane is less than the corresponding lane in other.

pub fn lanes_gt(self, other: Self) -> Mask16<LANES>

Test if each lane is greater than the corresponding lane in other.

pub fn lanes_le(self, other: Self) -> Mask16<LANES>

Test if each lane is less than or equal to the corresponding lane in other.

pub fn lanes_ge(self, other: Self) -> Mask16<LANES>

Test if each lane is greater than or equal to the corresponding lane in other.

impl<const LANES: usize> SimdI16<LANES> where
Self: LanesAtMost32,

pub fn saturating_add(self, second: Self) -> Self

Lanewise saturating add.

Examples

let x = SimdI16::from_array([MIN, 0, 1, MAX]);
let max = SimdI16::splat(MAX);
let unsat = x + max;
let sat = x.saturating_add(max);
assert_eq!(unsat, SimdI16::from_array([-1, MAX, MIN, -2]));
assert_eq!(sat, SimdI16::from_array([-1, MAX, MAX, MAX]));

pub fn saturating_sub(self, second: Self) -> Self

Lanewise saturating subtract.

Examples

let x = SimdI16::from_array([MIN, -2, -1, MAX]);
let max = SimdI16::splat(MAX);
let unsat = x - max;
let sat = x.saturating_sub(max);
assert_eq!(unsat, SimdI16::from_array([1, MAX, MIN, 0]));
assert_eq!(sat, SimdI16::from_array([MIN, MIN, MIN, 0]));

pub fn saturating_abs(self) -> Self

Lanewise saturating absolute value, implemented in Rust. As abs(), except the MIN value becomes MAX instead of itself.

Examples

use core_simd::*;

use core::i16::{MIN, MAX};

let x = SimdI16::splat([MIN, -2, 0, 3]); let unsat = x.abs(); let sat = x.saturating_abs(); assert_eq!(unsat, SimdI16::from_array([MIN, 2, 0, 3]); assert_eq!(sat, SimdI16::from_array([MAX, 2, 0, 3]));

pub fn saturating_neg(self) -> Self

Lanewise saturating negation, implemented in Rust. As neg(), except the MIN value becomes MAX instead of itself.

Examples

use core_simd::*;

use core::i16::{MIN, MAX};

let x = SimdI16::splat([MIN, -2, 3, MAX]); let unsat = -x; let sat = x.saturating_neg(); assert_eq!(unsat, SimdI16::from_array([MIN, 2, -3, MIN + 1]); assert_eq!(sat, SimdI16::from_array([MAX, 2, -3, MIN + 1]));

impl<const LANES: usize> SimdI16<LANES> where
Self: LanesAtMost32,

pub const fn splat(value: i16) -> Self

Construct a SIMD vector by setting all lanes to the given value.

pub const fn as_slice(&self) -> &[i16]

Returns a slice containing the entire SIMD vector.

pub fn as_mut_slice(&mut self) -> &mut [i16]

Returns a mutable slice containing the entire SIMD vector.

pub const fn from_array(array: [i16; LANES]) -> Self

Converts an array to a SIMD vector.

pub const fn to_array(self) -> [i16; LANES]

Converts a SIMD vector to an array.

impl SimdI16<2>

pub fn shuffle<const IDX: [u32; 2]>(self, second: Self) -> Self

A const SIMD shuffle that takes 2 SIMD vectors and produces another vector, using the indices in the const parameter. The first or “self” vector will have its lanes indexed from 0, and the second vector will have its first lane indexed at $n. Indices must be in-bounds of either vector at compile time.

Some SIMD shuffle instructions can be quite slow, so avoiding them by loading data into the desired patterns in advance is preferred, but shuffles are still faster than storing and reloading from memory.

impl SimdI16<4>

pub fn shuffle<const IDX: [u32; 4]>(self, second: Self) -> Self

A const SIMD shuffle that takes 2 SIMD vectors and produces another vector, using the indices in the const parameter. The first or “self” vector will have its lanes indexed from 0, and the second vector will have its first lane indexed at $n. Indices must be in-bounds of either vector at compile time.

Some SIMD shuffle instructions can be quite slow, so avoiding them by loading data into the desired patterns in advance is preferred, but shuffles are still faster than storing and reloading from memory.

impl SimdI16<8>

pub fn shuffle<const IDX: [u32; 8]>(self, second: Self) -> Self

A const SIMD shuffle that takes 2 SIMD vectors and produces another vector, using the indices in the const parameter. The first or “self” vector will have its lanes indexed from 0, and the second vector will have its first lane indexed at $n. Indices must be in-bounds of either vector at compile time.

Some SIMD shuffle instructions can be quite slow, so avoiding them by loading data into the desired patterns in advance is preferred, but shuffles are still faster than storing and reloading from memory.

impl SimdI16<16>

pub fn shuffle<const IDX: [u32; 16]>(self, second: Self) -> Self

A const SIMD shuffle that takes 2 SIMD vectors and produces another vector, using the indices in the const parameter. The first or “self” vector will have its lanes indexed from 0, and the second vector will have its first lane indexed at $n. Indices must be in-bounds of either vector at compile time.

Some SIMD shuffle instructions can be quite slow, so avoiding them by loading data into the desired patterns in advance is preferred, but shuffles are still faster than storing and reloading from memory.

impl SimdI16<32>

pub fn shuffle<const IDX: [u32; 32]>(self, second: Self) -> Self

A const SIMD shuffle that takes 2 SIMD vectors and produces another vector, using the indices in the const parameter. The first or “self” vector will have its lanes indexed from 0, and the second vector will have its first lane indexed at $n. Indices must be in-bounds of either vector at compile time.

Some SIMD shuffle instructions can be quite slow, so avoiding them by loading data into the desired patterns in advance is preferred, but shuffles are still faster than storing and reloading from memory.

impl<const LANES: usize> SimdI16<LANES> where
Self: LanesAtMost32,
SimdI16<LANES>: LanesAtMost32,

pub fn is_positive(self) -> Mask16<LANES>

Returns true for each positive lane and false if it is zero or negative.

pub fn is_negative(self) -> Mask16<LANES>

Returns true for each negative lane and false if it is zero or positive.