Struct core_simd::SimdI64

#[repr(simd)]
pub struct SimdI64<const LANES: usize>(_)
 where
    Self: LanesAtMost32;

A SIMD vector of containing LANES i64 values.

Implementations

impl<const LANES: usize> SimdI64<LANES> where
Self: ToBytes + LanesAtMost32,

pub fn to_ne_bytes(self) -> <Self as ToBytes>::Bytes

Return the memory representation of this integer as a byte array in native byte order.

pub fn from_ne_bytes(bytes: <Self as ToBytes>::Bytes) -> Self

Create a native endian integer value from its memory representation as a byte array in native endianness.

impl<const LANES: usize> SimdI64<LANES> where
SimdI64<LANES>: LanesAtMost32,
SimdI64<LANES>: LanesAtMost32,
Mask64<LANES>: Mask,

pub fn lanes_eq(self, other: Self) -> Mask64<LANES>

Test if each lane is equal to the corresponding lane in other.

pub fn lanes_ne(self, other: Self) -> Mask64<LANES>

Test if each lane is not equal to the corresponding lane in other.

pub fn lanes_lt(self, other: Self) -> Mask64<LANES>

Test if each lane is less than the corresponding lane in other.

pub fn lanes_gt(self, other: Self) -> Mask64<LANES>

Test if each lane is greater than the corresponding lane in other.

pub fn lanes_le(self, other: Self) -> Mask64<LANES>

Test if each lane is less than or equal to the corresponding lane in other.

pub fn lanes_ge(self, other: Self) -> Mask64<LANES>

Test if each lane is greater than or equal to the corresponding lane in other.

impl<const LANES: usize> SimdI64<LANES> where
Self: LanesAtMost32,

pub fn saturating_add(self, second: Self) -> Self

Lanewise saturating add.

Examples

let x = SimdI64::from_array([MIN, 0, 1, MAX]);
let max = SimdI64::splat(MAX);
let unsat = x + max;
let sat = x.saturating_add(max);
assert_eq!(unsat, SimdI64::from_array([-1, MAX, MIN, -2]));
assert_eq!(sat, SimdI64::from_array([-1, MAX, MAX, MAX]));

pub fn saturating_sub(self, second: Self) -> Self

Lanewise saturating subtract.

Examples

let x = SimdI64::from_array([MIN, -2, -1, MAX]);
let max = SimdI64::splat(MAX);
let unsat = x - max;
let sat = x.saturating_sub(max);
assert_eq!(unsat, SimdI64::from_array([1, MAX, MIN, 0]));
assert_eq!(sat, SimdI64::from_array([MIN, MIN, MIN, 0]));

pub fn abs(self) -> Self

Lanewise absolute value, implemented in Rust. Every lane becomes its absolute value.

Examples

let xs = SimdI64::from_array([MIN, MIN +1, -5, 0]);
assert_eq!(xs.abs(), SimdI64::from_array([MIN, MAX, 5, 0]));

pub fn saturating_abs(self) -> Self

Lanewise saturating absolute value, implemented in Rust. As abs(), except the MIN value becomes MAX instead of itself.

Examples

let xs = SimdI64::from_array([MIN, -2, 0, 3]);
let unsat = xs.abs();
let sat = xs.saturating_abs();
assert_eq!(unsat, SimdI64::from_array([MIN, 2, 0, 3]));
assert_eq!(sat, SimdI64::from_array([MAX, 2, 0, 3]));

pub fn saturating_neg(self) -> Self

Lanewise saturating negation, implemented in Rust. As neg(), except the MIN value becomes MAX instead of itself.

Examples

let x = SimdI64::from_array([MIN, -2, 3, MAX]);
let unsat = -x;
let sat = x.saturating_neg();
assert_eq!(unsat, SimdI64::from_array([MIN, 2, -3, MIN + 1]));
assert_eq!(sat, SimdI64::from_array([MAX, 2, -3, MIN + 1]));

impl<const LANES: usize> SimdI64<LANES> where
Self: LanesAtMost32,

pub const fn splat(value: i64) -> Self

Construct a SIMD vector by setting all lanes to the given value.

pub const fn as_slice(&self) -> &[i64]

Returns a slice containing the entire SIMD vector.

pub fn as_mut_slice(&mut self) -> &mut [i64]

Returns a mutable slice containing the entire SIMD vector.

pub const fn from_array(array: [i64; LANES]) -> Self

Converts an array to a SIMD vector.

pub const fn to_array(self) -> [i64; LANES]

Converts a SIMD vector to an array.

impl SimdI64<2>

pub fn shuffle<const IDX: [u32; 2]>(self, second: Self) -> Self

A const SIMD shuffle that takes 2 SIMD vectors and produces another vector, using the indices in the const parameter. The first or “self” vector will have its lanes indexed from 0, and the second vector will have its first lane indexed at $n. Indices must be in-bounds of either vector at compile time.

Some SIMD shuffle instructions can be quite slow, so avoiding them by loading data into the desired patterns in advance is preferred, but shuffles are still faster than storing and reloading from memory.

pub fn reverse(self) -> Self

Reverse the order of the lanes in the vector.

pub fn interleave(self, other: Self) -> (Self, Self)

Interleave two vectors.

Produces two vectors with lanes taken alternately from self and other.

The first result contains the first LANES / 2 lanes from self and other, alternating, starting with the first lane of self.

The second result contains the last LANES / 2 lanes from self and other, alternating, starting with the lane LANES / 2 from the start of self.

This particular permutation is efficient on many architectures.

let a = SimdU32::from_array([0, 1, 2, 3]);
let b = SimdU32::from_array([4, 5, 6, 7]);
let (x, y) = a.interleave(b);
assert_eq!(x.to_array(), [0, 4, 1, 5]);
assert_eq!(y.to_array(), [2, 6, 3, 7]);

pub fn deinterleave(self, other: Self) -> (Self, Self)

Deinterleave two vectors.

The first result takes every other lane of self and then other, starting with the first lane.

The second result takes every other lane of self and then other, starting with the second lane.

This particular permutation is efficient on many architectures.

let a = SimdU32::from_array([0, 4, 1, 5]);
let b = SimdU32::from_array([2, 6, 3, 7]);
let (x, y) = a.deinterleave(b);
assert_eq!(x.to_array(), [0, 1, 2, 3]);
assert_eq!(y.to_array(), [4, 5, 6, 7]);

impl SimdI64<4>

pub fn shuffle<const IDX: [u32; 4]>(self, second: Self) -> Self

A const SIMD shuffle that takes 2 SIMD vectors and produces another vector, using the indices in the const parameter. The first or “self” vector will have its lanes indexed from 0, and the second vector will have its first lane indexed at $n. Indices must be in-bounds of either vector at compile time.

Some SIMD shuffle instructions can be quite slow, so avoiding them by loading data into the desired patterns in advance is preferred, but shuffles are still faster than storing and reloading from memory.

pub fn reverse(self) -> Self

Reverse the order of the lanes in the vector.

pub fn interleave(self, other: Self) -> (Self, Self)

Interleave two vectors.

Produces two vectors with lanes taken alternately from self and other.

The first result contains the first LANES / 2 lanes from self and other, alternating, starting with the first lane of self.

The second result contains the last LANES / 2 lanes from self and other, alternating, starting with the lane LANES / 2 from the start of self.

This particular permutation is efficient on many architectures.

let a = SimdU32::from_array([0, 1, 2, 3]);
let b = SimdU32::from_array([4, 5, 6, 7]);
let (x, y) = a.interleave(b);
assert_eq!(x.to_array(), [0, 4, 1, 5]);
assert_eq!(y.to_array(), [2, 6, 3, 7]);

pub fn deinterleave(self, other: Self) -> (Self, Self)

Deinterleave two vectors.

The first result takes every other lane of self and then other, starting with the first lane.

The second result takes every other lane of self and then other, starting with the second lane.

This particular permutation is efficient on many architectures.

let a = SimdU32::from_array([0, 4, 1, 5]);
let b = SimdU32::from_array([2, 6, 3, 7]);
let (x, y) = a.deinterleave(b);
assert_eq!(x.to_array(), [0, 1, 2, 3]);
assert_eq!(y.to_array(), [4, 5, 6, 7]);

impl SimdI64<8>

pub fn shuffle<const IDX: [u32; 8]>(self, second: Self) -> Self

A const SIMD shuffle that takes 2 SIMD vectors and produces another vector, using the indices in the const parameter. The first or “self” vector will have its lanes indexed from 0, and the second vector will have its first lane indexed at $n. Indices must be in-bounds of either vector at compile time.

Some SIMD shuffle instructions can be quite slow, so avoiding them by loading data into the desired patterns in advance is preferred, but shuffles are still faster than storing and reloading from memory.

pub fn reverse(self) -> Self

Reverse the order of the lanes in the vector.

pub fn interleave(self, other: Self) -> (Self, Self)

Interleave two vectors.

Produces two vectors with lanes taken alternately from self and other.

The first result contains the first LANES / 2 lanes from self and other, alternating, starting with the first lane of self.

The second result contains the last LANES / 2 lanes from self and other, alternating, starting with the lane LANES / 2 from the start of self.

This particular permutation is efficient on many architectures.

let a = SimdU32::from_array([0, 1, 2, 3]);
let b = SimdU32::from_array([4, 5, 6, 7]);
let (x, y) = a.interleave(b);
assert_eq!(x.to_array(), [0, 4, 1, 5]);
assert_eq!(y.to_array(), [2, 6, 3, 7]);

pub fn deinterleave(self, other: Self) -> (Self, Self)

Deinterleave two vectors.

The first result takes every other lane of self and then other, starting with the first lane.

The second result takes every other lane of self and then other, starting with the second lane.

This particular permutation is efficient on many architectures.

let a = SimdU32::from_array([0, 4, 1, 5]);
let b = SimdU32::from_array([2, 6, 3, 7]);
let (x, y) = a.deinterleave(b);
assert_eq!(x.to_array(), [0, 1, 2, 3]);
assert_eq!(y.to_array(), [4, 5, 6, 7]);

impl SimdI64<16>

pub fn shuffle<const IDX: [u32; 16]>(self, second: Self) -> Self

A const SIMD shuffle that takes 2 SIMD vectors and produces another vector, using the indices in the const parameter. The first or “self” vector will have its lanes indexed from 0, and the second vector will have its first lane indexed at $n. Indices must be in-bounds of either vector at compile time.

Some SIMD shuffle instructions can be quite slow, so avoiding them by loading data into the desired patterns in advance is preferred, but shuffles are still faster than storing and reloading from memory.

pub fn reverse(self) -> Self

Reverse the order of the lanes in the vector.

pub fn interleave(self, other: Self) -> (Self, Self)

Interleave two vectors.

Produces two vectors with lanes taken alternately from self and other.

The first result contains the first LANES / 2 lanes from self and other, alternating, starting with the first lane of self.

The second result contains the last LANES / 2 lanes from self and other, alternating, starting with the lane LANES / 2 from the start of self.

This particular permutation is efficient on many architectures.

let a = SimdU32::from_array([0, 1, 2, 3]);
let b = SimdU32::from_array([4, 5, 6, 7]);
let (x, y) = a.interleave(b);
assert_eq!(x.to_array(), [0, 4, 1, 5]);
assert_eq!(y.to_array(), [2, 6, 3, 7]);

pub fn deinterleave(self, other: Self) -> (Self, Self)

Deinterleave two vectors.

The first result takes every other lane of self and then other, starting with the first lane.

The second result takes every other lane of self and then other, starting with the second lane.

This particular permutation is efficient on many architectures.

let a = SimdU32::from_array([0, 4, 1, 5]);
let b = SimdU32::from_array([2, 6, 3, 7]);
let (x, y) = a.deinterleave(b);
assert_eq!(x.to_array(), [0, 1, 2, 3]);
assert_eq!(y.to_array(), [4, 5, 6, 7]);

impl SimdI64<32>

pub fn shuffle<const IDX: [u32; 32]>(self, second: Self) -> Self

A const SIMD shuffle that takes 2 SIMD vectors and produces another vector, using the indices in the const parameter. The first or “self” vector will have its lanes indexed from 0, and the second vector will have its first lane indexed at $n. Indices must be in-bounds of either vector at compile time.

Some SIMD shuffle instructions can be quite slow, so avoiding them by loading data into the desired patterns in advance is preferred, but shuffles are still faster than storing and reloading from memory.

pub fn reverse(self) -> Self

Reverse the order of the lanes in the vector.

pub fn interleave(self, other: Self) -> (Self, Self)

Interleave two vectors.

Produces two vectors with lanes taken alternately from self and other.

The first result contains the first LANES / 2 lanes from self and other, alternating, starting with the first lane of self.

The second result contains the last LANES / 2 lanes from self and other, alternating, starting with the lane LANES / 2 from the start of self.

This particular permutation is efficient on many architectures.

let a = SimdU32::from_array([0, 1, 2, 3]);
let b = SimdU32::from_array([4, 5, 6, 7]);
let (x, y) = a.interleave(b);
assert_eq!(x.to_array(), [0, 4, 1, 5]);
assert_eq!(y.to_array(), [2, 6, 3, 7]);

pub fn deinterleave(self, other: Self) -> (Self, Self)

Deinterleave two vectors.

The first result takes every other lane of self and then other, starting with the first lane.

The second result takes every other lane of self and then other, starting with the second lane.

This particular permutation is efficient on many architectures.

let a = SimdU32::from_array([0, 4, 1, 5]);
let b = SimdU32::from_array([2, 6, 3, 7]);
let (x, y) = a.deinterleave(b);
assert_eq!(x.to_array(), [0, 1, 2, 3]);
assert_eq!(y.to_array(), [4, 5, 6, 7]);

impl<const LANES: usize> SimdI64<LANES> where
Self: LanesAtMost32,

pub fn horizontal_sum(self) -> i64

Horizontal wrapping add. Returns the sum of the lanes of the vector, with wrapping addition.

pub fn horizontal_product(self) -> i64

Horizontal wrapping multiply. Returns the product of the lanes of the vector, with wrapping multiplication.

pub fn horizontal_and(self) -> i64

Horizontal bitwise “and”. Returns the cumulative bitwise “and” across the lanes of the vector.

pub fn horizontal_or(self) -> i64

Horizontal bitwise “or”. Returns the cumulative bitwise “or” across the lanes of the vector.

pub fn horizontal_xor(self) -> i64

Horizontal bitwise “xor”. Returns the cumulative bitwise “xor” across the lanes of the vector.

pub fn horizontal_max(self) -> i64

Horizontal maximum. Returns the maximum lane in the vector.

pub fn horizontal_min(self) -> i64

Horizontal minimum. Returns the minimum lane in the vector.

impl<const LANES: usize> SimdI64<LANES> where
Self: LanesAtMost32,
SimdI64<LANES>: LanesAtMost32,
Mask64<LANES>: Mask,

pub fn is_positive(self) -> Mask64<LANES>

Returns true for each positive lane and false if it is zero or negative.

pub fn is_negative(self) -> Mask64<LANES>

Returns true for each negative lane and false if it is zero or positive.

pub fn signum(self) -> Self

Returns numbers representing the sign of each lane.

0 if the number is zero

1 if the number is positive

-1 if the number is negative