Struct core_simd::SimdMaskSize

#[repr(transparent)]
pub struct SimdMaskSize<const LANES: usize>(_)
 where
    SimdIsize<LANES>: LanesAtMost32;

A mask equivalent to SimdIsize, where all bits in the lane must be either set or unset.

Implementations

impl<const LANES: usize> SimdMaskSize<LANES> where
SimdIsize<LANES>: LanesAtMost32,

pub fn splat(value: bool) -> Self

Construct a mask by setting all lanes to the given value.

pub fn test(&self, lane: usize) -> bool

Tests the value of the specified lane.

Panics

Panics if lane is greater than or equal to the number of lanes in the vector.

pub fn set(&mut self, lane: usize, value: bool)

Sets the value of the specified lane.

Panics

Panics if lane is greater than or equal to the number of lanes in the vector.

pub unsafe fn from_int_unchecked(value: SimdIsize<LANES>) -> Self

Creates a mask from an integer vector.

Safety

All lanes must be either 0 or -1.

pub fn from_int(value: SimdIsize<LANES>) -> Self

Creates a mask from an integer vector.

Panics

Panics if any lane is not 0 or -1.