Struct core_simd::SimdU32

#[repr(simd)]pub struct SimdU32<const LANES: usize>(_);

A SIMD vector of containing LANES u32 values.

Implementations

impl<const LANES: usize> SimdU32<LANES>

pub fn lanes_eq(&self, other: &Self) -> Mask32<LANES>

Test if each lane is equal to the corresponding lane in other.

pub fn lanes_ne(&self, other: &Self) -> Mask32<LANES>

Test if each lane is not equal to the corresponding lane in other.

pub fn lanes_lt(&self, other: &Self) -> Mask32<LANES>

Test if each lane is less than the corresponding lane in other.

pub fn lanes_gt(&self, other: &Self) -> Mask32<LANES>

Test if each lane is greater than the corresponding lane in other.

pub fn lanes_le(&self, other: &Self) -> Mask32<LANES>

Test if each lane is less than or equal to the corresponding lane in other.

pub fn lanes_ge(&self, other: &Self) -> Mask32<LANES>

Test if each lane is greater than or equal to the corresponding lane in other.

impl<const LANES: usize> SimdU32<LANES>

pub const fn splat(value: u32) -> Self

Construct a SIMD vector by setting all lanes to the given value.

pub const fn as_slice(&self) -> &[u32]

Returns a slice containing the entire SIMD vector.

pub fn as_mut_slice(&mut self) -> &mut [u32]

Returns a mutable slice containing the entire SIMD vector.

pub const fn from_array(array: [u32; LANES]) -> Self

Converts an array to a SIMD vector.

pub const fn to_array(self) -> [u32; LANES]

Converts a SIMD vector to an array.