Struct core_simd::f64x2

#[repr(simd)]pub struct f64x2(_, _);

Vector of two f64 values

Implementations

impl f64x2

pub const fn splat(value: f64) -> Self

Construct a vector by setting all lanes to the given value.

pub const fn new(v0: f64, v1: f64) -> Self

Construct a vector by setting each lane to the given values.

impl f64x2

pub fn to_bits(self) -> u64x2

Raw transmutation to an unsigned integer vector type with the same size and number of lanes.

pub fn from_bits(bits: u64x2) -> Self

Raw transmutation from an unsigned integer vector type with the same size and number of lanes.

pub fn abs(self) -> Self

Produces a vector where every lane has the absolute value of the equivalently-indexed lane in self.

impl f64x2

#[must_use = "method returns a new vector and does not mutate the original value"]pub fn floor(self) -> Self

Returns the largest integer less than or equal to each lane.

#[must_use = "method returns a new vector and does not mutate the original value"]pub fn ceil(self) -> Self

Returns the smallest integer greater than or equal to each lane.

pub unsafe fn to_int_unchecked(self) -> i64x2

Rounds toward zero and converts to the same-width integer type, assuming that the value is finite and fits in that type.

Safety

The value must:

Not be NaN

Not be infinite

Be representable in the return type, after truncating off its fractional part

pub fn round_from_int(value: i64x2) -> Self

Creates a floating-point vector from an integer vector. Rounds values that are not exactly representable.