Struct core_simd::i64x8

#[repr(simd)]pub struct i64x8(_, _, _, _, _, _, _, _);

Vector of eight i64 values

Implementations

impl i64x8

pub const fn splat(value: i64) -> Self

Construct a vector by setting all lanes to the given value.

pub const fn new(
v0: i64,
v1: i64,
v2: i64,
v3: i64,
v4: i64,
v5: i64,
v6: i64,
v7: i64
) -> Self

Construct a vector by setting each lane to the given values.