Trait core_simd::MaskExt

pub trait MaskExt<Mask> {
    fn lanes_eq(&self, other: &Self) -> Mask;

    fn lanes_ne(&self, other: &Self) -> Mask;

    fn lanes_lt(&self, other: &Self) -> Mask;

    fn lanes_gt(&self, other: &Self) -> Mask;

    fn lanes_le(&self, other: &Self) -> Mask;

    fn lanes_ge(&self, other: &Self) -> Mask;
}

Mask-related operations using a particular mask layout.

Required methods

fn lanes_eq(&self, other: &Self) -> Mask

Test if each lane is equal to the corresponding lane in other.

fn lanes_ne(&self, other: &Self) -> Mask

Test if each lane is not equal to the corresponding lane in other.

fn lanes_lt(&self, other: &Self) -> Mask

Test if each lane is less than the corresponding lane in other.

fn lanes_gt(&self, other: &Self) -> Mask

Test if each lane is greater than the corresponding lane in other.

fn lanes_le(&self, other: &Self) -> Mask

Test if each lane is less than or equal to the corresponding lane in other.

fn lanes_ge(&self, other: &Self) -> Mask

Test if each lane is greater than or equal to the corresponding lane in other.

Loading content...