Trait core_simd::SimdArray

pub trait SimdArray<const LANES: usize>: LanesAtMost32 where
    SimdUsize<LANES>: LanesAtMost32,
    SimdIsize<LANES>: LanesAtMost32,
    MaskSize<LANES>: Mask,
    Self: Sized, {
    type Scalar: Copy + Sized;

    const LANES: usize;

    fn splat(val: Self::Scalar) -> Self;

    fn gather_or(
        slice: &[Self::Scalar], 
        idxs: SimdUsize<LANES>, 
        or: Self
    ) -> Self { ... }

    fn gather_or_default(slice: &[Self::Scalar], idxs: SimdUsize<LANES>) -> Self
    where
        Self::Scalar: Default,
    { ... }

    fn gather_select(
        slice: &[Self::Scalar], 
        mask: MaskSize<LANES>, 
        idxs: SimdUsize<LANES>, 
        or: Self
    ) -> Self { ... }

    fn scatter(self, slice: &mut [Self::Scalar], idxs: SimdUsize<LANES>) { ... }

    fn scatter_select(
        self, 
        slice: &mut [Self::Scalar], 
        mask: MaskSize<LANES>, 
        idxs: SimdUsize<LANES>
    ) { ... }
}

A representation of a vector as an “array” with indices, implementing operations applicable to any vector type based solely on “having lanes”, and describing relationships between vector and scalar types.

Associated Types

type Scalar: Copy + Sized

The scalar type in every lane of this vector type.

Associated Constants

const LANES: usize

The number of lanes for this vector.

Required methods

fn splat(val: Self::Scalar) -> Self

Generates a SIMD vector with the same value in every lane.

Provided methods

fn gather_or(slice: &[Self::Scalar], idxs: SimdUsize<LANES>, or: Self) -> Self

SIMD gather: construct a SIMD vector by reading from a slice, using potentially discontiguous indices. If an index is out of bounds, that lane instead selects the value from the “or” vector.

let vec: Vec<i32> = vec![10, 11, 12, 13, 14, 15, 16, 17, 18];
let idxs = SimdUsize::<4>::from_array([9, 3, 0, 5]);
let alt = SimdI32::from_array([-5, -4, -3, -2]);

let result = SimdI32::<4>::gather_or(&vec, idxs, alt); // Note the lane that is out-of-bounds.
assert_eq!(result, SimdI32::from_array([-5, 13, 10, 15]));

fn gather_or_default(slice: &[Self::Scalar], idxs: SimdUsize<LANES>) -> Self where
Self::Scalar: Default,

SIMD gather: construct a SIMD vector by reading from a slice, using potentially discontiguous indices. Out-of-bounds indices instead use the default value for that lane (0).

let vec: Vec<i32> = vec![10, 11, 12, 13, 14, 15, 16, 17, 18];
let idxs = SimdUsize::<4>::from_array([9, 3, 0, 5]);

let result = SimdI32::<4>::gather_or_default(&vec, idxs); // Note the lane that is out-of-bounds.
assert_eq!(result, SimdI32::from_array([0, 13, 10, 15]));

fn gather_select(
slice: &[Self::Scalar],
mask: MaskSize<LANES>,
idxs: SimdUsize<LANES>,
or: Self
) -> Self

SIMD gather: construct a SIMD vector by reading from a slice, using potentially discontiguous indices. Out-of-bounds or masked indices instead select the value from the “or” vector.

let vec: Vec<i32> = vec![10, 11, 12, 13, 14, 15, 16, 17, 18];
let idxs = SimdUsize::<4>::from_array([9, 3, 0, 5]);
let alt = SimdI32::from_array([-5, -4, -3, -2]);
let mask = MaskSize::from_array([true, true, true, false]); // Note the mask of the last lane.

let result = SimdI32::<4>::gather_select(&vec, mask, idxs, alt); // Note the lane that is out-of-bounds.
assert_eq!(result, SimdI32::from_array([-5, 13, 10, -2]));

fn scatter(self, slice: &mut [Self::Scalar], idxs: SimdUsize<LANES>)

SIMD scatter: write a SIMD vector’s values into a slice, using potentially discontiguous indices. Out-of-bounds indices are not written. scatter writes “in order”, so if an index receives two writes, only the last is guaranteed.

let mut vec: Vec<i32> = vec![10, 11, 12, 13, 14, 15, 16, 17, 18];
let idxs = SimdUsize::<4>::from_array([9, 3, 0, 0]);
let vals = SimdI32::from_array([-27, 82, -41, 124]);

vals.scatter(&mut vec, idxs); // index 0 receives two writes.
assert_eq!(vec, vec![124, 11, 12, 82, 14, 15, 16, 17, 18]);

fn scatter_select(
self,
slice: &mut [Self::Scalar],
mask: MaskSize<LANES>,
idxs: SimdUsize<LANES>
)

SIMD scatter: write a SIMD vector’s values into a slice, using potentially discontiguous indices. Out-of-bounds or masked indices are not written. scatter_select writes “in order”, so if an index receives two writes, only the last is guaranteed.

let mut vec: Vec<i32> = vec![10, 11, 12, 13, 14, 15, 16, 17, 18];
let idxs = SimdUsize::<4>::from_array([9, 3, 0, 0]);
let vals = SimdI32::from_array([-27, 82, -41, 124]);
let mask = MaskSize::from_array([true, true, true, false]); // Note the mask of the last lane.

vals.scatter_select(&mut vec, mask, idxs); // index 0's second write is masked, thus omitted.
assert_eq!(vec, vec![-41, 11, 12, 82, 14, 15, 16, 17, 18]);