Trait core_simd::SimdFloat

pub trait SimdFloat: Sized + Sealed {
    fn simd_min(self, other: Self) -> Self;
    fn simd_max(self, other: Self) -> Self;
    fn simd_clamp(self, min: Self, max: Self) -> Self;
}

🔬 This is a nightly-only experimental API. (portable_simd)

SIMD operations on vectors of floating point numbers.

Required Methods

fn simd_min(self, other: Self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Returns the minimum of each lane.

If one of the values is NAN, then the other value is returned.

fn simd_max(self, other: Self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Returns the maximum of each lane.

If one of the values is NAN, then the other value is returned.

fn simd_clamp(self, min: Self, max: Self) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Restrict each lane to a certain interval unless it is NaN.

For each lane in self, returns the corresponding lane in max if the lane is greater than max, and the corresponding lane in min if the lane is less than min. Otherwise returns the lane in self.