Trait core_simd::SimdOrd

pub trait SimdOrd: SimdPartialOrd {
    fn simd_max(self, other: Self) -> Self;
    fn simd_min(self, other: Self) -> Self;
    fn simd_clamp(self, min: Self, max: Self) -> Self;
}

🔬This is a nightly-only experimental API. (portable_simd)

Parallel Ord.

Required Methods

fn simd_max(self, other: Self) -> Self

🔬This is a nightly-only experimental API. (portable_simd)

Returns the lane-wise maximum with other.

fn simd_min(self, other: Self) -> Self

🔬This is a nightly-only experimental API. (portable_simd)

Returns the lane-wise minimum with other.

fn simd_clamp(self, min: Self, max: Self) -> Self

🔬This is a nightly-only experimental API. (portable_simd)

Restrict each lane to a certain interval.

For each lane, returns max if self is greater than max, and min if self is less than min. Otherwise returns self.

Panics

Panics if min > max on any lane.