Trait core_simd::Vector

pub trait Vector: Sealed {
    type Scalar: Copy + Sized;

    const LANES: usize;

    fn splat(val: Self::Scalar) -> Self;
}

🔬 This is a nightly-only experimental API. (portable_simd)

A representation of a vector as an “array” with indices, implementing operations applicable to any vector type based solely on “having lanes”, and describing relationships between vector and scalar types.

Associated Types

type Scalar: Copy + Sized

🔬 This is a nightly-only experimental API. (portable_simd)

The scalar type in every lane of this vector type.

Associated Constants

const LANES: usize

🔬 This is a nightly-only experimental API. (portable_simd)

The number of lanes for this vector.

Required methods

fn splat(val: Self::Scalar) -> Self

🔬 This is a nightly-only experimental API. (portable_simd)

Generates a SIMD vector with the same value in every lane.