Trait core_simd::simd::SimdPartialEq

pub trait SimdPartialEq {
    type Mask;

    fn simd_eq(self, other: Self) -> Self::Mask;
    fn simd_ne(self, other: Self) -> Self::Mask;
}

๐ฌThis is a nightly-only experimental API. (portable_simd)

Parallel PartialEq.

Required Associated Types

type Mask

๐ฌThis is a nightly-only experimental API. (portable_simd)

The mask type returned by each comparison.

Required Methods

fn simd_eq(self, other: Self) -> Self::Mask

๐ฌThis is a nightly-only experimental API. (portable_simd)

Test if each lane is equal to the corresponding lane in other.

fn simd_ne(self, other: Self) -> Self::Mask

๐ฌThis is a nightly-only experimental API. (portable_simd)

Test if each lane is equal to the corresponding lane in other.