Trait core_simd::simd::SimdPartialOrd

pub trait SimdPartialOrd: SimdPartialEq {
    fn simd_lt(self, other: Self) -> Self::Mask;
    fn simd_le(self, other: Self) -> Self::Mask;
    fn simd_gt(self, other: Self) -> Self::Mask;
    fn simd_ge(self, other: Self) -> Self::Mask;
}

🔬 This is a nightly-only experimental API. (portable_simd)

Parallel PartialOrd.

Required Methods

fn simd_lt(self, other: Self) -> Self::Mask

🔬 This is a nightly-only experimental API. (portable_simd)

Test if each lane is less than the corresponding lane in other.

fn simd_le(self, other: Self) -> Self::Mask

🔬 This is a nightly-only experimental API. (portable_simd)

Test if each lane is less than or equal to the corresponding lane in other.

fn simd_gt(self, other: Self) -> Self::Mask

🔬 This is a nightly-only experimental API. (portable_simd)

Test if each lane is greater than the corresponding lane in other.

fn simd_ge(self, other: Self) -> Self::Mask

🔬 This is a nightly-only experimental API. (portable_simd)

Test if each lane is greater than or equal to the corresponding lane in other.