Trait core_simd::simd::SupportedLaneCount

pub trait SupportedLaneCount: Sealed { }

🔬 This is a nightly-only experimental API. (portable_simd)

Statically guarantees that a lane count is marked as supported.

This trait is sealed: the list of implementors below is total. Users do not have the ability to mark additional LaneCount<N> values as supported. Only SIMD vectors with supported lane counts are constructable.